@@ -34,65 +34,68 @@
     "license": "MIT",
     "dependencies": {
         "@entrylabs/event": "^1.0.3",
-        "@types/node": "^11.9.5",
         "autoprefixer": "^9.4.3",
-        "axios": "^0.18.0",
         "events": "^3.0.0",
+        "fontfaceonload": "^1.0.2",
+        "file-saver": "^2.0.2",
         "hammerjs": "^2.0.8",
         "html-webpack-plugin": "^3.2.0",
         "html-webpack-template": "^6.2.0",
-        "hyperid": "^1.4.1",
+        "isomorphic-fetch": "^2.2.1",
         "jest": "^24.1.0",
-        "lodash": "^4.17.11",
+        "lodash": "^4.17.15",
         "postcss-flexbugs-fixes": "^4.1.0",
         "simplebar": "^3.1.1",
         "uglifyjs-webpack-plugin": "^2.1.2",
-        "webpack": "^4.28.3"
+        "webpack": "^4.28.3",
+        "xss-filters": "^1.2.7"
     },
     "devDependencies": {
-        "@babel/core": "7.0.0",
-        "@babel/plugin-proposal-class-properties": "^7.0.0",
-        "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-        "@babel/plugin-syntax-dynamic-import": "7.0.0",
-        "@babel/plugin-transform-modules-commonjs": "^7.0.0",
-        "@babel/plugin-transform-object-assign": "7.0.0",
-        "@babel/plugin-transform-runtime": "7.0.0",
-        "@babel/preset-env": "7.0.0",
-        "@babel/runtime": "^7.0.0",
+        "@babel/core": "7.2.0",
+        "@babel/plugin-proposal-class-properties": "^7.2.0",
+        "@babel/plugin-proposal-private-methods": "^7.2.0",
+        "@babel/plugin-proposal-object-rest-spread": "^7.2.0",
+        "@babel/plugin-syntax-dynamic-import": "7.2.0",
+        "@babel/plugin-transform-modules-commonjs": "^7.2.0",
+        "@babel/plugin-transform-object-assign": "7.2.0",
+        "@babel/plugin-transform-runtime": "7.2.0",
+        "@babel/preset-env": "7.2.0",
+        "@babel/runtime": "^7.2.0",
         "@egjs/visible": "^2.1.0",
+        "@types/lodash": "^4.14.136",
+        "@types/node": "^11.9.5",
         "@types/pixi.js": "^4.8.6",
+        "@types/webpack-env": "^1.14.0",
+        "@typescript-eslint/parser": "^2.2.0",
         "awesome-typescript-loader": "^5.2.1",
-        "babel-eslint": "^8.2.2",
+        "babel-eslint": "^10.0.3",
         "babel-loader": "^8.0.0-beta",
         "babel-plugin-dynamic-import-node": "^2.0.0",
         "chai": "^3.2.0",
         "clean-webpack-plugin": "^1.0.1",
         "cross-env": "^5.1.4",
-        "css-loader": "^0.28.10",
-        "cssnano": "^4.1.7",
-        "eslint": "^4.18.2",
-        "eslint-config-prettier": "^4.1.0",
-        "eslint-loader": "^2.0.0",
-        "eslint-plugin-prettier": "^3.0.1",
+        "css-loader": "^3.2.0",
+        "cssnano": "^4.1.10",
+        "eslint": "^6.3.0",
+        "eslint-config-prettier": "^6.2.0",
+        "eslint-loader": "^3.0.0",
+        "eslint-plugin-prettier": "^3.1.0",
         "extract-text-webpack-plugin": "^4.0.0-beta.0",
-        "file-loader": "^1.1.11",
+        "file-loader": "^4.2.0",
         "hard-source-webpack-plugin": "^0.13.1",
         "jest-jenkins-reporter": "^1.0.2",
-        "karma": "^4.0.0",
+        "karma": "^4.2.0",
         "karma-chai": "^0.1.0",
-        "karma-chrome-launcher": "^0.2.0",
-        "karma-mocha": "^0.2.0",
-        "less": "^3.9.0",
-        "less-loader": "^4.0.6",
+        "karma-chrome-launcher": "^3.1.0",
+        "karma-mocha": "^1.3.0",
+        "less": "^3.10.1",
+        "less-loader": "^5.0.0",
         "mini-css-extract-plugin": "^0.5.0",
-        "mocha": "^2.5.3",
-        "postcss-cssnext": "^3.1.0",
-        "postcss-import": "^12.0.1",
+        "mocha": "^6.2.0",
         "postcss-loader": "^3.0.0",
-        "prettier": "^1.16.4",
-        "prettier-eslint": "^8.8.2",
+        "prettier": "^1.18.2",
+        "prettier-eslint": "^9.0.0",
         "style-loader": "^0.20.2",
-        "sugarss": "^2.0.0",
         "typescript": "^3.2.2",
         "url-loader": "^1.0.1",
         "webpack-cli": "^3.2.1",
