@@ -472,37 +472,36 @@ Lang.Blocks = {
     "behaviorConduct03013": "교통사고",
     "behaviorConduct03014": "승강기 안전사고",
     "behaviorConduct03015": "미세먼지",
-    "behaviorConduct01001002": "태풍이 오기 전",
-    "behaviorConduct01001003": "태풍 주의보 시",
-    "behaviorConduct01001004": "태풍 경보 시",
-    "behaviorConduct01001005": "태풍이 지나간 후",
-    "behaviorConduct01002002": "홍수 예보 시",
-    "behaviorConduct01002003": "물이 밀려들 때",
-    "behaviorConduct01002004": "물이 빠진 후",
-    "behaviorConduct01003002": "호우 예보 시",
-    "behaviorConduct01003003": "호우 주의보∙경보 시",
-    "behaviorConduct01003004": "호우가 지나간 후",
-    "behaviorConduct01004002": "강풍이 오기 전",
-    "behaviorConduct01004003": "강풍 발생 시",
-    "behaviorConduct01004004": "강풍이 지나간 후",
-    "behaviorConduct01005002": "대설 발생 시",
-    "behaviorConduct01006002": "한파 발생 시",
+    "behaviorConduct01001001": "태풍이 오기 전",
+    "behaviorConduct01001002": "태풍이 왔을 때",
+    "behaviorConduct01001003": "태풍이 지나간 후",
+    "behaviorConduct01002002": "홍수가 오기 전",
+    "behaviorConduct01002003": "홍수가 났을 때",
+    "behaviorConduct01002004": "홍수가 지나간 후",
+    "behaviorConduct01003002": "큰 비가 오기 전",
+    "behaviorConduct01003003": "큰 비가 내릴 때",
+    "behaviorConduct01003004": "큰 비가 그친 후",
+    "behaviorConduct01004001": "강풍이 오기 전",
+    "behaviorConduct01004002": "강풍 발생 시",
+    "behaviorConduct01005002": "대설이 오기 전",
+    "behaviorConduct01005003": "대설 내일 때",
+    "behaviorConduct01005004": "눈이 그친 후",
+    "behaviorConduct01006001": "한파가 오기 전",
+    "behaviorConduct01006002": "한파 왔을 때",
+    "behaviorConduct01007001": "풍랑이 험할 때",
     "behaviorConduct01007002": "풍랑이 오기 전",
-    "behaviorConduct01007003": "풍랑 주의보·경보 시",
-    "behaviorConduct01008002": "황사 예보 시",
-    "behaviorConduct01008003": "황사 발생 시",
-    "behaviorConduct01008004": "황사가 지나간 후",
+    "behaviorConduct01008001": "황사가 오기 전",
+    "behaviorConduct01008002": "황사가 왔을 때",
+    "behaviorConduct01008003": "황사가 지나간 후",
     "behaviorConduct01009001": "폭염이 오기 전",
-    "behaviorConduct01009002": "폭염 주의보 시",
-    "behaviorConduct01009003": "폭염 경보 시",
-    "behaviorConduct01010001": "가뭄이 오기 전",
+    "behaviorConduct01009002": "폭염이 왔을 때",
+    "behaviorConduct01010001": "가뭄이 왔을 때",
     "behaviorConduct01011009": "지진이 오기 전",
-    "behaviorConduct01011010": "지진 발생시",
+    "behaviorConduct01011010": "지진이 왔을 때",
     "behaviorConduct01011012": "지진 대피 후",
     "behaviorConduct01013002": "해일이 오기 전",
-    "behaviorConduct01013003": "해일 발생 시",
-    "behaviorConduct01014001": "산사태 주의보 시",
-    "behaviorConduct01014002": "산사태 경보 시",
+    "behaviorConduct01013003": "해일이 왔을 때",
+    "behaviorConduct01014001": "산사태가 오기 전",
     "behaviorConduct03002002": "화상 처치",
     "behaviorConduct03002003": "고열 예방",
     "behaviorConduct03002004": "뱀에 물렸을 때 처치",
@@ -1733,7 +1732,7 @@ Lang.Blocks = {
     "CALC_calc_operation_tan": "탄젠트값",
     "CALC_calc_operation_floor": "소수점 버림값",
     "CALC_calc_operation_ceil": "소수점 올림값",
-    "CALC_calc_operation_round": "반올림값",
+    "CALC_calc_operation_round": "소수점 반올림값",
     "CALC_calc_operation_factorial": "펙토리얼값",
     "CALC_calc_operation_asin": "아크사인값",
     "CALC_calc_operation_acos": "아크코사인값",
@@ -2193,7 +2192,9 @@ Lang.Buttons = {
     "multi_line": "여러 줄 쓰기",
     "lecture_finish_popup_reset_lecture": "처음부터 다시하기",
     "lecture_finish_popup_show_output": "실습한 코드보기",
-    "default_hint_title": "어떻게 만들까?"
+    "lecture_view_hints": "잘 모르겠어요!",
+    "default_hint_title": "어떻게 만들까?",
+    "show_only_vector": "벡터 모아보기"
 };
 Lang.Menus = {
     "nothing_to_import": "불러오기 할 항목이 없습니다.",
@@ -2274,7 +2275,7 @@ Lang.Menus = {
     "delete_discuss_file": "파일을 삭제하시겠습니까?",
     "discuss_save_question": "글을 저장하시겠습니까?",
     "discuss_cancle_question": "작성을 취소하시겠습니까?",
-    "discuss_saved": "이 저장되었습니다.",
+    "discuss_saved": "이(가) 저장되었습니다.",
     "discuss_no_write_permission": "현재 로그인된 계정으로는 글을 작성하실 수 없습니다.",
     "discuss_no_project_permission": "현재 로그인된 계정으로는 작품을 게시하실 수 없습니다.",
     "discuss_write_abuse_detected": "짧은 시간안에 여러 글이 작성되었습니다.\n1분 뒤에 다시 시도해주세요.",
@@ -2550,13 +2551,18 @@ Lang.Menus = {
     "draw_new_ques_2": "그리기 화면으로 이동하시겠습니까?",
     "draw_new_go": "이동하기",
     "draw_new_stay": "이동하지 않기",
+    "file_upload_warn_title_image": "아래와 같은 그림은 이용약관 및 관련 법률에 의해 제재를 받으실 수 있습니다.",
+    "file_upload_warn_desc_image": "폭력적이고 잔인한 그림<br>선정적인 신체노출 그림<br>불쾌감을 주거나 혐오감을 일으키는 그림<br>무단 사용이 금지된 저작권의 그림",
+    "file_upload_warn_title_sound": "아래와 같은 소리는 이용약관 및 관련 법률에 의해 제재를 받을 수 있습니다.",
+    "file_upload_warn_desc_sound": "폭력적이고 잔인한 소리<br>선정적인 내용의 소리<br>불쾌감을 주거나 혐오감을 일으키는 소리<br>무단 사용이 금지된 저작권의 소리",
+    "file_upload_warn_link": "저작권에 대해 알아보기",
     "file_upload_desc_1": "아래와 같은 그림은 이용약관 및 관련 법률에 의해 제재를 받으실 수 있습니다.",
     "file_upload_desc_2": "폭력적이고 잔인한 그림",
     "file_upload_desc_3": "선정적인 신체노출 그림",
     "file_upload_desc_4": "불쾌감을 주거나 혐오감을 일으키는 그림",
     "file_upload_desc_5": "* 위와 같은 내용은 이용약관 및 관련 법률에 의해 제재를 받으실 수 있습니다.",
-    "picture_upload_warn_1": "10MB 이하의 jpg, png, bmp 형식의 파일을 추가할 수 있습니다.",
-    "sprite_upload_warn": "10MB 이하의 jpg, png, bmp 이미지 또는 eo 형식의 오브젝트를 추가할 수 있습니다.",
+    "picture_upload_warn_1": "10MB 이하의 jpg, png, bmp, svg 형식의 파일을 추가할 수 있습니다.\njpg, png, bmp 형식은 비트맵 모드, svg 형식은 벡터 모드로 편집됩니다.",
+    "sprite_upload_warn": "10MB 이하의 jpg, png, bmp, svg 이미지 또는 eo 형식의 오브젝트를 추가할 수 있습니다.",
     "sound_upload_warn_1": "10MB 이하의 mp3 형식의 파일을 추가할 수 있습니다.",
     "sound_edit_warn": "기다려 주세요!<br>새로운 소리 편집 기능을 연구중입니다.",
     "file_upload_loading": "로딩중",
@@ -2646,6 +2652,16 @@ Lang.Menus = {
     "lecture_finish_popup_title": "학습 다시하기",
     "lecture_finish_popup_content_1": "실습완료한 단계를<br>처음부터 다시 실습하시겠습니까?",
     "lecture_finish_popup_content_2": "* 다시 실습해도 실습 완료 상태는 유지됩니다.",
+    "lecture_start_title": "학습 시작",
+    "lecture_goal_title": "학습 목표",
+    "lecture_goal_content": "무엇을 배울지 알아볼까요?",
+    "lecture_prepare_title": "학습 준비",
+    "lecture_summary_title": "학습 정리",
+    "lecture_summary_content": "배운 내용을 정리해 봅시다.",
+    "lecture_quiz_title": "확인 문제",
+    "lecture_quiz_content": "확인 문제를 통해 배운 내용을 점검해 봅시다.",
+    "lecture_quiz_check_answer": "정답확인",
+    "default_hint_title": "어떻게 만들까?",
     "only_pdf": ".pdf형식의 파일만 입력 가능합니다.",
     "enter_project_video": "적어도 하나의 작품이나 영상을 입력하세요.",
     "enter_title": "제목을 입력하세요.",
@@ -3439,7 +3455,7 @@ Lang.Menus = {
     "basic_mission_content": "강아지 로봇을 만드는 공장에서 우연한 정전으로 혼자서 생각할 수 있게 된 엔트리봇! 공장을 탈출하고 자유를 찾을 수 있도록 엔트리봇을 도와주세요!",
     "application_mission": "응용미션: 엔트리봇 우주여행",
     "write_article": "글쓰기",
-    "write_contest": "응모하기",
+    "write_contest": "제출하기",
     "view_all_articles": "모든 글 보기",
     "view_own_articles": "내가 쓴 글 보기",
     "view_qna_articles": "답변 없는 글 보기",
@@ -4229,6 +4245,8 @@ Lang.Menus = {
     "kumsungMid_edu_entry": "금성출판사 엔트리 학습하기",
     "gilbutEdu_edu_entry": "길벗출판사 엔트리 학습하기",
     "miraenCoding_edu_entry": "미래엔 엔트리 학습하기",
+    "textbook_help_data": "도움자료",
+    "textbook_teaching_activity": "수업 교안&활동지",
     "textbook_kyohacksa": "교학사",
     "textbook_kumsung": "금성출판사",
     "textbook_donga": "동아출판",
@@ -4526,6 +4544,16 @@ Lang.Menus = {
     "group_add_course_3": "추가하고 싶은 강의 모음을 관심 강의 모음으로 등록해 주세요.",
     "group_add_course_4": "강의 모음 보기",
     "hw_main_title": "프로그램 다운로드",
+    "painter_contest_title": "엔트리 캠페인 #2 개성있는 작품 만들기",
+    "painter_content_head": "확 바뀐 그림판과 새로운 오브젝트",
+    "painter_content_body1": "벡터 이미지 편집과 새로 추가된 오브젝트로",
+    "painter_content_body2": "더 멋진 작품 만드는 법",
+    "painter_content_button": "캠페인 보기",
+    "campaign_contest_title": "아름다운 인터넷 세상을 만들어요",
+    "campaign_content_head": "엔트리 작품 공모전",
+    "campaign_content_body1": "'인터넷윤리'에 대해 알기 쉽고 재미있게",
+    "campaign_content_body2": "작품을 만들어 응모해주세요!",
+    "campaign_content_button": "자세히 보기",
     "code_management_1": "엔트리 캠페인 #1 코드 잘 관리하기",
     "code_management_2": "나의 보관함과 메모 활용",
     "code_management_3": "새 기능 나의 보관함과 메모로 하는 코드 관리 비법!",
@@ -4673,11 +4701,12 @@ Lang.Menus = {
     "search_미디어아트": "미디어 아트",
     "search_피지컬": "피지컬",
     "search_기타": "기타",
-    "discuss_write_textarea_placeholer": "엔트리 커뮤니티는 우리가 함께 만들어가는 공간입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!",
-    "discuss_write_textarea_placeholer_qna": "여기는 묻고답하기 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!",
-    "discuss_write_textarea_placeholer_tips": "여기는 노하우&팁 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!",
-    "discuss_write_textarea_placeholer_free": "여기는 엔트리 이야기 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!",
-    "discuss_write_textarea_placeholer_report": "여기는 제안 및 건의 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!",
+    "discuss_write_textarea_placeholer": "엔트리 커뮤니티는 우리가 함께 만들어가는 공간입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!\n\n 욕설이나 음란물과 같은 부적절한 내용의 글을 작성할 경우 이용정지 될 수 있습니다.",
+    "discuss_write_textarea_placeholer_qna": "여기는 묻고답하기 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!\n\n 욕설이나 음란물과 같은 부적절한 내용의 글을 작성할 경우 이용정지 될 수 있습니다.",
+    "discuss_write_textarea_placeholer_tips": "여기는 노하우&팁 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!\n\n 욕설이나 음란물과 같은 부적절한 내용의 글을 작성할 경우 이용정지 될 수 있습니다.",
+    "discuss_write_textarea_placeholer_free": "여기는 엔트리 이야기 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!\n\n 욕설이나 음란물과 같은 부적절한 내용의 글을 작성할 경우 이용정지 될 수 있습니다.",
+    "discuss_write_textarea_placeholer_report": "여기는 제안 및 건의 게시판입니다.\n서로 배려하는 말로 즐거운 커뮤니티를 만들어요!\n\n 욕설이나 음란물과 같은 부적절한 내용의 글을 작성할 경우 이용정지 될 수 있습니다.",
+    "discuss_write_textarea_placeholer_contest2019": "※ 아래 양식으로 작성하여 제출해주세요.\n\n1. 학년 (예: 초등학교 5학년)\n\n2. 작품 URL (‘공유하기 > 작품 공유하기’에 작품을 공유한 다음, 작품 URL을 복사하여 적어주세요.)\n\n3. 작품 설명\n\n※ 공모전 수상자에게는 이메일이 발송될 예정입니다.\n엔트리의 ‘내 정보 수정‘에 이메일 정보를 꼭 입력해주세요!",
     "maze_road": "길",
     "account_deletion": "회원탈퇴",
     "bug_report_too_many_request": "신고 내용이 전송 되고 있습니다. 잠시 후에 다시 시도해주시길 바랍니다.",
@@ -4808,12 +4837,12 @@ Lang.Msgs = {
     "select_hardware_button": "하드웨어 선택하기",
     "select_the_expansion_block": "수업에 활용할 확장 블록을 선택하세요.",
     "select_expansion_block_button": "확장 블록 선택하기",
-    "expansion_weather_description": "기온, 강수량, 미세먼지 농도 등 한국의 날씨와 관련된 블록들의 모음입니다.",
-    "expansion_festival_description": "국내 지역별 다양한 행사 및 축제와 관련된 블록들의 모음입니다",
-    "expansion_translate_description": "파파고를 이용하여 다른 언어로 번역할 수 있는 블록들의 모음입니다.",
-    "expansion_behaviorConductDisaster_description": "자연재난 발생 시 국민이 지켜야하는 기본적인 행동요령에 대한 블록들의 모음입니다. (국민안전처 제공)",
-    "expansion_behaviorConductLifeSafety_description": "생활속에서 안전을 위해 국민이 지켜야하는 기본적인 행동요령에 대한 블록들의 모음입니다. (국민안전처 제공)",
-    "expansion_tts_description": "nVoice를 이용하여 다양한 목소리로 문장을 읽을 수 있는 블록들의 모음입니다."
+    "expansion_weather_description": "기온, 강수량, 미세먼지 농도 등 한국의 날씨와 관련된 블록 모음입니다. [웨더아이 제공]",
+    "expansion_festival_description": "국내 지역별 다양한 행사 및 축제와 관련된 불록 모음 입니다. [한국관광공사 제공]",
+    "expansion_translate_description": "파파고를 이용하여 다른 언어로 번역할 수 있는 블록 모음입니다. [네이버 제공]",
+    "expansion_behaviorConductDisaster_description": "자연재난 발생 시 국민이 지켜야하는 기본적인 행동요령에 대한 블록들의 모음입니다. [국민안전처 제공]",
+    "expansion_behaviorConductLifeSafety_description": "생활속 안전을 위해 국민이 지켜야하는 행동요령에 대한 블록 모음입니다. [국민안전처 제공]",
+    "expansion_tts_description": "nVoice 음성합성 기술로 다양한 목소리로 문장을 읽는 블록모음 입니다. [네이버 제공]"
 };
 Lang.Users = {
     "auth_failed": "인증에 실패하였습니다",
@@ -4861,6 +4890,9 @@ Lang.Users = {
     "not_valid_email": "올바른 형식의 이메일 주소를 입력해주세요."
 };
 Lang.Workspace = {
+    "picture_rotation": "회전",
+    "picture_size": "크기",
+    "thickness": "굵기",
     "follow": "따라해보기",
     "SaveWithPicture": "저장되지 않은 그림이 있습니다. 저장하시겠습니까?",
     "RecursiveCallWarningTitle": "함수 호출 제한",
@@ -4878,6 +4910,7 @@ Lang.Workspace = {
     "block_helper": "블록 도움말",
     "new_project": "새 프로젝트",
     "add_object": "오브젝트 추가하기",
+    "download_image_name": "엔트리 블록",
     "all": "전체",
     "animal": "동물",
     "arduino_entry": "아두이노 연결 프로그램",
@@ -4888,7 +4921,7 @@ Lang.Workspace = {
     "cannot_add_picture": "실행중에는 모양을 추가할 수 없습니다.",
     "cannot_add_sound": "실행중에는 소리를 추가할 수 없습니다.",
     "cannot_add_expansion_block": "실행중에는 확장 블럭을 추가할 수 없습니다.",
-    "cannot_edit_click_to_stop": "실행중에는 수정할 수 없습니다.<br>클릭하여 정지하기.",
+    "cannot_edit_click_to_stop": "실행 중에는 수정할 수 없습니다.<br>클릭하여 정지하기.",
     "cannot_open_private_project": "비공개 작품은 불러올 수 없습니다. 홈으로 이동합니다.",
     "cannot_save_running_project": "실행 중에는 저장할 수 없습니다.",
     "character_gen": "캐릭터 만들기",
@@ -5038,7 +5071,7 @@ Lang.Workspace = {
     "Variable_use_all_objects": "모든 오브젝트에서 사용",
     "Variable_use_this_object": "이 오브젝트에서 사용",
     "Variable_used_at_all_objects": "모든 오브젝트에서 사용되는 변수",
-    "Variable_create_cloud": "공유 변수로 사용 <br>(서버에 저장됩니다)",
+    "Variable_create_cloud": "공유 변수로 사용 \n(서버에 저장됩니다)",
     "Variable_used_at_special_object": "특정 오브젝트에서만 사용되는 변수",
     "Variable_used_objects": "사용된 오브젝트",
     "draw_new": "새로 그리기",
@@ -5049,29 +5082,6 @@ Lang.Workspace = {
     "painter_edit": "편집 ▼",
     "get_file": "가져오기",
     "copy_file": "복사하기",
-    "cut_picture": "자르기",
-    "paste_picture": "붙이기",
-    "remove_all": "모두 지우기",
-    "new_picture": "새그림",
-    "picture_size": "크기",
-    "picture_rotation": "회전",
-    "flip_horizontal": "좌우뒤집기",
-    "flip_vertical": "상하뒤집기",
-    "thickness": "굵기",
-    "regular": "보통",
-    "bold": "굵게",
-    "font_underline": "밑줄",
-    "font_tilt": "기울임",
-    "font_cancel": "취소선",
-    "font_color": "글꼴 색상",
-    "font_fill": "채우기 색상",
-    "font_family": "글꼴",
-    "align_left": "왼쪽 정렬",
-    "align_center": "가운데 정렬",
-    "align_right": "오른쪽 정렬",
-    "italic": "기울임",
-    "textStyle": "글자",
-    "add_picture": "모양 추가",
     "select_picture": "모양 선택",
     "select_sound": "소리 선택",
     "Size": "크기",
@@ -5101,7 +5111,7 @@ Lang.Workspace = {
     "robot_guide": "로봇 연결 안내",
     "python_guide": "엔트리파이선 이용 안내",
     "show_list_workspace": "리스트 노출",
-    "List_create_cloud": "공유 리스트로 사용 <br>(서버에 저장됩니다)",
+    "List_create_cloud": "공유 리스트로 사용 \n(서버에 저장됩니다)",
     "confirm_quit": "바꾼 내용을 저장하지 않았습니다.",
     "confirm_load_temporary": "저장되지 않은 작품이 있습니다. 여시겠습니까?",
     "login_to_save": "로그인후에 저장 바랍니다.",
@@ -5178,7 +5188,9 @@ Lang.Workspace = {
     "brightness": "명도",
     "spoid": "스포이드",
     "palette_mode": "팔레트 모드",
-    "slider_mode": "슬라이더 모드"
+    "slider_mode": "슬라이더 모드",
+    "toast_error_title_object_center": "주의",
+    "toast_error_contents_object_center": "오브젝트 중심점이 오브젝트와 멀리 떨어져 실행화면 밖으로 벗어날 경우, 원위치로 돌아옵니다."
 };
 Lang.code = "코드보기";
 Lang.EntryStatic = {
@@ -5301,6 +5313,7 @@ Lang.EntryStatic = {
     "general": "일반",
     "curriculum_is_open_true": "공개",
     "curriculum_open_false": "비공개",
+    "contest2019": "공모전",
     "exhibit": "공모전",
     "notice": "공지사항",
     "qna": "묻고답하기",
@@ -5443,7 +5456,7 @@ Lang.Helper = {
     "get_rotation_direction": "해당 오브젝트의 방향값, 이동 방향값을 의미합니다.",
     "calc_share": "앞 수에서 뒤 수를 나누어 생긴 몫을 의미합니다.",
     "calc_mod": "앞 수에서 뒤 수를 나누어 생긴 나머지를 의미합니다.",
-    "calc_operation": "입력한 수에 대한 다양한 수학식의 계산값입니다. (제곱, 루트, 사인값, 코사인값, 탄젠트값, 아크사인값, 아크코사인값, 아크탄젠트값, 로그값, 자연로그값, 소수점 부분, 소수점 버림값, 소수점 올림값, 반올림값, 펙토리얼값, 절대값)",
+    "calc_operation": "입력한 수에 대한 다양한 수학식의 계산값입니다. (제곱, 루트, 사인값, 코사인값, 탄젠트값, 아크사인값, 아크코사인값, 아크탄젠트값, 로그값, 자연로그값, 소수점 부분, 소수점 버림값, 소수점 올림값, 소수점 반올림값, 펙토리얼값, 절대값)",
     "get_date": "현재 연도, 월, 일, 시각과 같이 시간에 대한 값입니다.",
     "distance_something": "해당 오브젝트와 선택한 오브젝트 또는 마우스 포인터 간의 거릿값입니다.",
     "get_sound_duration": "선택한 소리의 길이(초) 값입니다.",
@@ -5901,7 +5914,8 @@ Lang.Helper = {
     "count_lifeSafety_behavior": "선택한 생활 속 문제가 발생했을 때 해야할 행동요령의 수입니다.<br> *옵션<br> 생활 속 문제: 응급처치, 심폐소생술, 소화기·소화전 사용법, 식중독, 산행안전, 실종유괴, 성폭력, 학교폭력, 가정 폭력, 억류 및 납치, 교통사고, 승강기 안전사고, 미세먼지<br> 상황: 생활 속 문제에 따라 다름 ",
     "get_lifeSafety_behavior": "선택한 생활 속 문제가 발생했을 때 해야할 행동요령에 대한 정보입니다.<br> *옵션<br> 생활 속 문제: 응급처치, 심폐소생술, 소화기·소화전 사용법, 식중독, 산행안전, 실종유괴, 성폭력, 학교폭력, 가정 폭력, 억류 및 납치, 교통사고, 승강기 안전사고, 미세먼지<br> 상황: 생활 속 문제에 따라 다름 ",
     "read_text": "입력한 문자값을 설정된 목소리로 읽습니다.<br>입력은 2500자까지 가능합니다.<br>인터넷에 연결되어 있지 않거나 인터넷 환경이 불안할 경우, 해당 블록이 실행되지 않고 다음 블록으로 넘어갈 수 있습니다.",
-    "set_tts_property": "선택한 목소리가 선택한 속도와 선택한 음높이로 설정됩니다."
+    "set_tts_property": "선택한 목소리가 선택한 속도와 선택한 음높이로 설정됩니다.",
+    "read_text_wait_with_block": "입력한 문자값을 읽어준 후 다음 블록을 실행합니다."
 };
 Lang.Category = {
     "entrybot_friends": "엔트리봇 친구들",
@@ -5940,6 +5954,9 @@ Lang.Category = {
     "stuff_others": "기타",
     "fantasy": "판타지",
     "interface": "인터페이스",
+    "interface_website": "웹사이트",
+    "interface_game": "게임",
+    "interface_others": "기타",
     "background": "배경",
     "background_outdoor": "실외",
     "background_indoor": "실내",
@@ -6033,7 +6050,14 @@ Lang.Fonts = {
     "yisunshin": "이순신체",
     "goyang": "고양체",
     "flower_road": "꽃길체",
-    "uhbeemysen": "어비마이센체"
+    "uhbeemysen": "어비마이센체",
+    "sd_comic_stencil": "산돌 코믹스탠실",
+    "sd_childfundkorea": "산돌 초록우산 어린이",
+    "sd_cinema_theater": "산돌 씨네마극장",
+    "sd_mapssi": "산돌 맵씨",
+    "sd_shabang": "산돌 별이샤방샤방",
+    "sd_woodcarving": "산돌 목각",
+    "sd_yongbi": "산돌 용비어천가"
 };
 Lang.Hw = {
     "note": "음표",
@@ -6551,6 +6575,7 @@ Lang.template = {
     "tts_title": "%1",
     "read_text": "%1 읽어주기 %2",
     "set_tts_property": "%1 목소리를 %2 속도 %3 음높이로 설정하기 %4",
+    "read_text_wait_with_block": "%1 읽어주고 기다리기 %2",
     "load_expansion_block": "확장 블록 불러오기",
     "expansion_block_descriptions": "확장 블록은 인터넷이 연결되어 있어야 정상적으로 동작합니다.",
     "weather_title_text": "날씨",
@@ -7588,7 +7613,7 @@ Lang.PythonHelper = {
     "quotient_and_mod_exampleCode": "def when_start():\n    Entry.print_for_sec(10 // 3, 2)\n    Entry.print_for_sec(10 % 3, 2)",
     "quotient_and_mod_exampleDesc": "[시작하기]버튼을 클릭하면 10 / 3의 몫인 3을 2초 동안 말하고, 나머지인 1을 2초 동안 말합니다.",
     "calc_operation_desc": "A의 연산값입니다.",
-    "calc_operation_elements": "A, B-- 연산하고자 하는 숫자값<br>① A ** 2 : A를 제곱한 값<br>② math.sqrt(A): A의 루트값<br>③ math.sin(A): A의 사인값<br>④ math.cos(A): A의 코사인 값<br>⑤ math.tan(A):  A의 탄젠트값 <br>⑥ math.asin(A): A의 아크사인값<br>⑦ math.acos(A): A의 아크코사인값<br>⑧ math.atan(): A의 아크탄젠트값<br>⑨ math.log10(A): A의 로그값<br>⑩ math.log(A): A의 자연로그값<br>⑪ A - math.floor(A): A의 소수점 부분<br>⑫ math.floor(A): A의 소수점 버림값<br>⑬ math.ceil(A): A의 소수점 올림값<br>⑭ math.round(A): A의 반올림값<br>⑮ math.factorial(A): A의 팩토리얼 값<br>⑯ math.fabs(A): A의 절댓값",
+    "calc_operation_elements": "A, B-- 연산하고자 하는 숫자값<br>① A ** 2 : A를 제곱한 값<br>② math.sqrt(A): A의 루트값<br>③ math.sin(A): A의 사인값<br>④ math.cos(A): A의 코사인 값<br>⑤ math.tan(A):  A의 탄젠트값 <br>⑥ math.asin(A): A의 아크사인값<br>⑦ math.acos(A): A의 아크코사인값<br>⑧ math.atan(): A의 아크탄젠트값<br>⑨ math.log10(A): A의 로그값<br>⑩ math.log(A): A의 자연로그값<br>⑪ A - math.floor(A): A의 소수점 부분<br>⑫ math.floor(A): A의 소수점 버림값<br>⑬ math.ceil(A): A의 소수점 올림값<br>⑭ math.round(A): A의 소수점 반올림값<br>⑮ math.factorial(A): A의 팩토리얼 값<br>⑯ math.fabs(A): A의 절댓값",
     "calc_operation_exampleCode": "def when_start():\n    Entry.print_for_sec(10 ** 2, 2)\n    Entry.print_for_sec(math.sqrt(9), 2)\n    Entry.print_for_sec(math.sin(90), 2)\n    Entry.print_for_sec(math.fabs(-10), 2)",
     "calc_operation_exampleDesc": "[시작하기]버튼을 클릭하면 10의 제곱, 9의 루트값, 90의 사인값, -10의 절댓값을 각 2초 동안 말합니다.",
     "get_project_timer_value_desc": "이 명령이 실행되는 순간 초시계에 저장된 값입니다.",
@@ -7763,12 +7788,51 @@ Lang.Painter = {
     "font": "글꼴",
     "font_color": "글꼴 색상",
     "outline_color": "윤곽선 색상",
+    "outline_thickness": "윤곽선 굵기",
     "fill_color": "채우기 색상",
     "inversion": "반전",
     "font_size": "글꼴 크기",
     "font_family": "글꼴 스타일",
     "fullscreen": "전체화면",
-    "exit_fullscreen": "전체화면 해제"
+    "exit_fullscreen": "전체화면 해제",
+    "picture_import": "모양 가져오기",
+    "edit_way": "편집 방법",
+    "line": "직선",
+    "vertex": "꼭지점",
+    "bitmap": "비트맵",
+    "vector": "벡터",
+    "forward": "앞으로",
+    "backward": "뒤로",
+    "front": "맨 앞으로",
+    "back": "맨 뒤로",
+    "group": "그룹화",
+    "ungroup": "그룹 해제",
+    "copy_picture": "복사",
+    "delete_picture": "삭제",
+    "cut_picture": "자르기",
+    "paste_picture": "붙여넣기",
+    "remove_all": "모두 삭제",
+    "new_picture": "새그림",
+    "edit": "편집",
+    "picture_size": "크기",
+    "picture_rotation": "회전",
+    "thickness": "굵기",
+    "font_fill": "채우기 색상",
+    "painter_file_save": "저장하기",
+    "painter_file_saveas": "새 모양으로 저장",
+    "file": "파일",
+    "Pencil": "펜",
+    "Line": "직선",
+    "Path": "형태",
+    "Pan": "이동",
+    "Select": "선택",
+    "Rectangle": "사각형",
+    "Ellipse": "원",
+    "Text": "글상자",
+    "Fill": "채우기",
+    "Eraser": "지우기",
+    "get_file": "가져오기",
+    "copy_file": "복사하기"
 };
 
 Lang.Support = ["ko", "en", "vn", "jp"];
