@@ -472,29 +472,29 @@ Lang.Blocks = {
     "behaviorConduct03013": "交通事故",
     "behaviorConduct03014": "エレベーター事故",
     "behaviorConduct03015": "浮遊粒子状物質",
-    "behaviorConduct01001002": "台風が来る前",
-    "behaviorConduct01001003": "台風注意報が出された時",
-    "behaviorConduct01001004": "台風警報が出された時",
-    "behaviorConduct01001005": "台風が通り過ぎた後",
+    "behaviorConduct01001001": "台風が来る前",
+    "behaviorConduct01001002": "台風警報が出された時",
+    "behaviorConduct01001003": "台風が通り過ぎた後",
     "behaviorConduct01002002": "洪水予報時",
     "behaviorConduct01002003": "水が押し寄せる時",
     "behaviorConduct01002004": "水が抜けた後",
     "behaviorConduct01003002": "大雨予報時",
     "behaviorConduct01003003": "大雨注意報∙警報が出された時",
     "behaviorConduct01003004": "大雨が過ぎ去った後",
-    "behaviorConduct01004002": "強風が来る前",
-    "behaviorConduct01004003": "強風発生時",
-    "behaviorConduct01004004": "強風が過ぎ去った後",
-    "behaviorConduct01005002": "大雪発生時",
+    "behaviorConduct01004001": "強風が来る前",
+    "behaviorConduct01004002": "強風発生時",
+    "behaviorConduct01005002": "大雪が来る前",
+    "behaviorConduct01005003": "大雪発生時",
+    "behaviorConduct01005004": "大雪が通り過ぎた後",
+    "behaviorConduct01006001": "寒波が押し寄せる前",
     "behaviorConduct01006002": "寒波発生時",
+    "behaviorConduct01007001": "波浪注意報・警報時",
     "behaviorConduct01007002": "波浪が押し寄せる前",
-    "behaviorConduct01007003": "波浪注意報・警報時",
-    "behaviorConduct01008002": "黄砂予報時",
-    "behaviorConduct01008003": "黄砂発生時",
-    "behaviorConduct01008004": "黄砂が過ぎ去った後",
+    "behaviorConduct01008001": "黄砂予報時",
+    "behaviorConduct01008002": "黄砂発生時",
+    "behaviorConduct01008003": "黄砂が過ぎ去った後",
     "behaviorConduct01009001": "猛暑が来る前",
     "behaviorConduct01009002": "猛暑注意報時",
-    "behaviorConduct01009003": "猛暑警報時",
     "behaviorConduct01010001": "干ばつが来る前",
     "behaviorConduct01011009": "地震が来る前",
     "behaviorConduct01011010": "地震発生時",
@@ -502,7 +502,6 @@ Lang.Blocks = {
     "behaviorConduct01013002": "津波が来る前",
     "behaviorConduct01013003": "津波発生時",
     "behaviorConduct01014001": "地滑り注意報時",
-    "behaviorConduct01014002": "地滑り警報時",
     "behaviorConduct03002002": "やけどの処置",
     "behaviorConduct03002003": "高熱の予防",
     "behaviorConduct03002004": "ヘビにかまれたときの処置",
@@ -975,7 +974,7 @@ Lang.Blocks = {
     "ARDUINO_convert_scale_5": "~",
     "ARDUINO_convert_scale_6": "",
     "ARDUINO_off": "Off",
-    "EXPANSION": "関数",
+    "EXPANSION": "拡張",
     "brightness": "明度",
     "BRUSH": "筆",
     "BRUSH_brush_erase_all": "すべての筆のあとを消す",
@@ -2193,7 +2192,9 @@ Lang.Buttons = {
     "multi_line": "改行して書く",
     "lecture_finish_popup_reset_lecture": "처음부터 다시하기",
     "lecture_finish_popup_show_output": "실습한 코드보기",
-    "default_hint_title": "どのように作るか"
+    "lecture_view_hints": "View hints",
+    "default_hint_title": "どのように作るか",
+    "show_only_vector": "Only Vector"
 };
 Lang.Menus = {
     "nothing_to_import": "読み込みできる項目がありません。",
@@ -2550,13 +2551,18 @@ Lang.Menus = {
     "draw_new_ques_2": "描く画面に移動しますか",
     "draw_new_go": "移動する",
     "draw_new_stay": "移動しない",
+    "file_upload_warn_title_image": "下のような画像は、利用規約および関連の法律によって制裁を受けることがあります。",
+    "file_upload_warn_desc_image": "暴力的で残酷な画像<br>性的に体の部位が露出している画像<br>不快感を与えたり、嫌悪感を与える画像<br>無断使用が禁止された画像",
+    "file_upload_warn_title_sound": "下記のようなサウンドは、利用規約および関連の法律によって制裁措置される場合があります。",
+    "file_upload_warn_desc_sound": "暴力的で残酷なサウンド<br>性的なサウンド<br>不快感や嫌悪感を与えるサウンド<br>無断使用が禁止されたサウンド",
+    "file_upload_warn_link": "著作権について",
     "file_upload_desc_1": "下のような画像は、利用規約および関連の法律によって制裁を受けることがあります。",
     "file_upload_desc_2": "暴力的で残酷な画像",
     "file_upload_desc_3": "性的に体の部位が露出している画像",
     "file_upload_desc_4": "不快感を与えたり、嫌悪感を与える画像",
     "file_upload_desc_5": "* 上のような内容は、利用規約および関連法律により制裁を受けることがあります。",
-    "picture_upload_warn_1": "10MB以下のjpg, png, bmp形式のファイルを追加できます。",
-    "sprite_upload_warn": "You can only upload image or object files such as jpg, png, bmp, eo that are less than 10MB.",
+    "picture_upload_warn_1": "10MB以下のjpg,png,bmp,svg形式の画像ファイルのみアップロードできます。\njpg,png,bmp形式はビットマップモードで,svg形式はベクターモードで編集されます。",
+    "sprite_upload_warn": "10MB以下のjpg,png,bmp,svg形式の画像やeo形式のオブジェクトファイルのみアップロードできます。",
     "sound_upload_warn_1": "10MB以下のmp3形式のファイルを追加できます。",
     "sound_edit_warn": "お待ちください。<br>ただいま、新しい音の編集機能を研究中です。",
     "file_upload_loading": "ローディング中",
@@ -2646,6 +2652,16 @@ Lang.Menus = {
     "lecture_finish_popup_title": "학습 다시하기",
     "lecture_finish_popup_content_1": "실습완료한 단계를<br>처음부터 다시 실습하시겠습니까?",
     "lecture_finish_popup_content_2": "* 다시 실습해도 실습 완료 상태는 유지됩니다.",
+    "lecture_start_title": "Start",
+    "lecture_goal_title": "Goal",
+    "lecture_goal_content": "Goal",
+    "lecture_prepare_title": "Prepare",
+    "lecture_summary_title": "Summary",
+    "lecture_summary_content": "Let's summarize what we learned.",
+    "lecture_quiz_title": "Quiz",
+    "lecture_quiz_content": "Let's check what we learned through the verification problem.",
+    "lecture_quiz_check_answer": "Check the answer",
+    "default_hint_title": "どのように作るか",
     "only_pdf": ".pdf形式のファイルのみ入力できます。",
     "enter_project_video": "最低1件の作品、または映像を入力してください。",
     "enter_title": "タイトルを入力してください。",
@@ -3439,7 +3455,7 @@ Lang.Menus = {
     "basic_mission_content": "ロボット犬を作る工場で偶然停電が起き、1人で考えることができるようになったエントリーボット! 工場を脱出して自由の身になれるよう、エントリーボットを手伝ってあげてください!",
     "application_mission": "応用ミッション: エントリーボットの宇宙旅行",
     "write_article": "投稿する",
-    "write_contest": "Apply",
+    "write_contest": "Submit",
     "view_all_articles": "すべての投稿を見る",
     "view_own_articles": "自分の投稿を見る",
     "view_qna_articles": "レスのないスレッド一覧",
@@ -4229,6 +4245,8 @@ Lang.Menus = {
     "kumsungMid_edu_entry": "ギムソン出版社 エントリーを学習する",
     "gilbutEdu_edu_entry": "ギルバート出版社 エントリーを学習する",
     "miraenCoding_edu_entry": "ミライエン エントリーを学習する",
+    "textbook_help_data": "助力資料",
+    "textbook_teaching_activity": "授業教案&活動誌",
     "textbook_kyohacksa": "Kyohaksa",
     "textbook_kumsung": "Kumsung Publishing",
     "textbook_donga": "Dong-a Publishing",
@@ -4526,6 +4544,16 @@ Lang.Menus = {
     "group_add_course_3": "追加したい講義集を関心講義集に登録してください。",
     "group_add_course_4": "講義集を見る",
     "hw_main_title": "プログラムダウンロード",
+    "painter_contest_title": "個性のある作品作り",
+    "painter_content_head": "新しく変わったペイント",
+    "painter_content_body1": "ベクトルイメージ編集と新しいオブジェクトを",
+    "painter_content_body2": "用いて素晴らしい作品を作る",
+    "painter_content_button": "キャンペーンを見る",
+    "campaign_contest_title": "Let's make a beautiful internet world",
+    "campaign_content_head": "Entry Project Contest",
+    "campaign_content_body1": "Create an originative project about 'internet ethics'",
+    "campaign_content_body2": "and apply for the contest!",
+    "campaign_content_button": "View details",
     "code_management_1": "ENTRYキャンペーン#1 コードを上手に管理しよう",
     "code_management_2": "マイボックスとメモの活用",
     "code_management_3": "新しい機能と一緒にコードを管理し、",
@@ -4678,6 +4706,7 @@ Lang.Menus = {
     "discuss_write_textarea_placeholer_tips": "このページはノウハウ＆ティップの掲示板です。\nお互い思いやる言葉を使って楽しいコミュニティにしましょう。",
     "discuss_write_textarea_placeholer_free": "このページはENTRYストーリー掲示板です。\nお互い思いやる言葉を使って楽しいコミュニティにしましょう。",
     "discuss_write_textarea_placeholer_report": "このページは提案＆意見掲示板です。\nお互い思いやる言葉を使って楽しいコミュニティにしましょう。",
+    "discuss_write_textarea_placeholer_contest2019": "※ 아래 양식으로 작성하여 제출해주세요.\n\n1. 학년 (예: 초등학교 5학년)\n\n2. 작품 URL (‘공유하기 > 작품 공유하기’에 작품을 공유한 다음, 작품 URL을 복사하여 적어주세요.)\n\n3. 작품 설명\n\n※ 공모전 수상자에게는 이메일이 발송될 예정입니다.\n엔트리의 ‘내 정보 수정‘에 이메일 정보를 꼭 입력해주세요!",
     "maze_road": "道",
     "account_deletion": "会員退会",
     "bug_report_too_many_request": "通報内容を送信しています。しばらくしてからやり直してください。",
@@ -4861,6 +4890,9 @@ Lang.Users = {
     "not_valid_email": "Please enter a correct type of email."
 };
 Lang.Workspace = {
+    "picture_rotation": "回転",
+    "picture_size": "サイズ",
+    "thickness": "太さ",
     "follow": "やってみる",
     "SaveWithPicture": "保存されていない図があります。保存しますか?",
     "RecursiveCallWarningTitle": "関数呼び出し制限",
@@ -4878,6 +4910,7 @@ Lang.Workspace = {
     "block_helper": "ブロックヘルプ",
     "new_project": "新しいプロジェクト",
     "add_object": "オブジェクトを追加する",
+    "download_image_name": "entry_ブロック",
     "all": "すべて",
     "animal": "動物",
     "arduino_entry": "Program for connect to Arduino",
@@ -5038,7 +5071,7 @@ Lang.Workspace = {
     "Variable_use_all_objects": "すべてのオブジェクトで使用",
     "Variable_use_this_object": "このオブジェクトで使用",
     "Variable_used_at_all_objects": "すべてのオブジェクトで使われている変数",
-    "Variable_create_cloud": "共有変数として使用 <br>(サーバーに保存されます)",
+    "Variable_create_cloud": "共有変数として使用 \n(サーバーに保存されます)",
     "Variable_used_at_special_object": "特定のオブジェクトでのみ使われている変数",
     "Variable_used_objects": "使用されたオブジェクト",
     "draw_new": "新しく描画",
@@ -5049,29 +5082,6 @@ Lang.Workspace = {
     "painter_edit": "編集 ▼",
     "get_file": "インポート",
     "copy_file": "コピーする",
-    "cut_picture": "切り取り",
-    "paste_picture": "貼り付け",
-    "remove_all": "すべて削除",
-    "new_picture": "新しい図",
-    "picture_size": "サイズ",
-    "picture_rotation": "回転",
-    "flip_horizontal": "左右反転",
-    "flip_vertical": "上下反転",
-    "thickness": "太さ",
-    "regular": "普通",
-    "bold": "太字",
-    "font_underline": "下線",
-    "font_tilt": "斜線",
-    "font_cancel": "取り消し線",
-    "font_color": "文字の色",
-    "font_fill": "背景色",
-    "font_family": "フォント",
-    "align_left": "左揃え",
-    "align_center": "中央揃え",
-    "align_right": "右揃え",
-    "italic": "斜体",
-    "textStyle": "文字",
-    "add_picture": "模様追加",
     "select_picture": "模様選択",
     "select_sound": "サウンド選択",
     "Size": "サイズ",
@@ -5101,7 +5111,7 @@ Lang.Workspace = {
     "robot_guide": "ロボット連結案内",
     "python_guide": "エントリーパイソン利用案内",
     "show_list_workspace": "リスト表示",
-    "List_create_cloud": "共有リストとして使用<br>(サーバーに保存されます)",
+    "List_create_cloud": "共有リストとして使用\n(サーバーに保存されます)",
     "confirm_quit": "変更した内容を保存しませんでした。",
     "confirm_load_temporary": "保存されていない作品があります。開きますか?",
     "login_to_save": "ログイン後に保存してください。",
@@ -5178,7 +5188,9 @@ Lang.Workspace = {
     "brightness": "明度",
     "spoid": "スポイト",
     "palette_mode": "パレットモード",
-    "slider_mode": "スライダーモード"
+    "slider_mode": "スライダーモード",
+    "toast_error_title_object_center": "注意！",
+    "toast_error_contents_object_center": "キャラクターの中心点がキャラクターから離れすぎため、修正しました。"
 };
 Lang.code = "コードを見る";
 Lang.EntryStatic = {
@@ -5301,6 +5313,7 @@ Lang.EntryStatic = {
     "general": "一般",
     "curriculum_is_open_true": "公開",
     "curriculum_open_false": "非公開",
+    "contest2019": "Contest",
     "exhibit": "Contest",
     "notice": "お知らせ",
     "qna": "Q＆A",
@@ -5901,7 +5914,8 @@ Lang.Helper = {
     "count_lifeSafety_behavior": "<br>選択した、生活トラブルが生じた時に取るべき行動要領の数です。",
     "get_lifeSafety_behavior": "<br>選択した、生活トラブルが生じた時に取るべき行動要領についての情報です。",
     "read_text": "入力した文字列を設定した音声で読み上げます。",
-    "set_tts_property": "選択した音声が選択した速度と選択したピッチで設定されます。"
+    "set_tts_property": "選択した音声が選択した速度と選択したピッチで設定されます。",
+    "read_text_wait_with_block": "入力した文字列を設定した音声で読み上げし、読み上げが終わると次のブロックを実行します。"
 };
 Lang.Category = {
     "entrybot_friends": "エントリーボットの友達",
@@ -5940,6 +5954,9 @@ Lang.Category = {
     "stuff_others": "その他",
     "fantasy": "ファンタジー",
     "interface": "インターフェイス",
+    "interface_website": "ウェブサイト",
+    "interface_game": "ゲーム",
+    "interface_others": "その他",
     "background": "背景",
     "background_outdoor": "室外",
     "background_indoor": "室内",
@@ -6021,19 +6038,26 @@ Lang.Fonts = {
     "jeju_hallasan": "漢拏山体",
     "gothic_coding": "コーディングゴシック体",
     "batang": "パタン体",
-    "gothic": "ゴシック体",
-    "myeongjo": "明朝体",
-    "pen_script": "筆記体",
-    "square_round": "コーディングゴシック体",
-    "jalnan": "コーディングゴシック体",
-    "designhouse": "コーディングゴシック体",
-    "malssami815": "コーディングゴシック体",
-    "dunggeunmo": "コーディングゴシック体",
-    "hs_bombaram": "コーディングゴシック体",
-    "yisunshin": "コーディングゴシック体",
-    "goyang": "コーディングゴシック体",
-    "flower_road": "コーディングゴシック体",
-    "uhbeemysen": "コーディングゴシック体"
+    "gothic": "ナヌムゴシック体",
+    "myeongjo": "ナヌム明朝",
+    "pen_script": "ナヌム手書き",
+    "square_round": "ナヌム角丸体",
+    "jalnan": "ジャルナン体",
+    "designhouse": "デザインハウス体",
+    "malssami815": "マルサミ８１５体",
+    "dunggeunmo": "丸文字体",
+    "hs_bombaram": "HSボムバラム体",
+    "yisunshin": "イスンシン体",
+    "goyang": "ゴヤン体",
+    "flower_road": "サンサンフラワーロード体",
+    "uhbeemysen": "アビーマイセン体",
+    "sd_comic_stencil": "Sandoll コミックステンシル",
+    "sd_childfundkorea": "Sandoll コドモノタメニ",
+    "sd_cinema_theater": "Sandoll シネマゲキジョウ",
+    "sd_mapssi": "Sandoll メッシ",
+    "sd_shabang": "Sandoll ホシガキラキラ",
+    "sd_woodcarving": "Sandoll キボリ",
+    "sd_yongbi": "Sandoll ヨンビ"
 };
 Lang.Hw = {
     "note": "音符",
@@ -6551,6 +6575,7 @@ Lang.template = {
     "tts_title": "%1",
     "read_text": "%1 を読み上げる %2",
     "set_tts_property": "%1 音声で速さを%2、ピッチを%3に設定する。%4",
+    "read_text_wait_with_block": "%1 を読み上げて待つ %2",
     "load_expansion_block": "追加する",
     "expansion_block_descriptions": "拡張ブロックが正常に動作するには、インターネット接続が必要です。",
     "weather_title_text": "天気",
@@ -6558,7 +6583,7 @@ Lang.template = {
     "festival_title_text": "イベント",
     "behaviorConductDisaster_title_text": "自然災害",
     "behaviorConductLifeSafety_title_text": "生活安全",
-    "tts_title_text": "を読み上げる",
+    "tts_title_text": "音声読み上げ",
     "hamster_hand_found": "hand found?",
     "hamster_value": "%1",
     "hamster_move_forward_once": "move forward once on board %1",
@@ -7763,12 +7788,51 @@ Lang.Painter = {
     "font": "フォント",
     "font_color": "文字の色",
     "outline_color": "輪郭線の色",
+    "outline_thickness": "枠線の太さ",
     "fill_color": "塗りつぶしの色",
     "inversion": "反転",
     "font_size": "サイズ",
     "font_family": "スタイル",
     "fullscreen": "全画面",
-    "exit_fullscreen": "全画面を終了"
+    "exit_fullscreen": "全画面を終了",
+    "picture_import": "模様インポート",
+    "edit_way": "編集方法",
+    "line": "直線",
+    "vertex": "頂点",
+    "bitmap": "ビットマップ",
+    "vector": "ベクター",
+    "forward": "前に",
+    "backward": "後ろに",
+    "front": "一番前に",
+    "back": "一番後ろに",
+    "group": "ｸﾞﾙｰﾌﾟ",
+    "ungroup": "ｸﾞﾙｰﾌﾟ 解除",
+    "copy_picture": "コピー",
+    "delete_picture": "削除",
+    "cut_picture": "切り取り",
+    "paste_picture": "貼り付け",
+    "remove_all": "全て削除",
+    "new_picture": "新しい図",
+    "edit": "編集",
+    "picture_size": "サイズ",
+    "picture_rotation": "回転",
+    "thickness": "太さ",
+    "font_fill": "背景色",
+    "painter_file_save": "保存する",
+    "painter_file_saveas": "新しい模様で保存",
+    "file": "ファイル",
+    "Pencil": "ペン",
+    "Line": "直線",
+    "Path": "形を変える",
+    "Pan": "移動",
+    "Select": "選択",
+    "Rectangle": "四角形",
+    "Ellipse": "円",
+    "Text": "テキストボックス",
+    "Fill": "塗りつぶし",
+    "Eraser": "クリア",
+    "get_file": "インポート",
+    "copy_file": "コピーする"
 };
 
 Lang.Support = ["ko", "en", "vn", "jp"];
