@@ -472,37 +472,36 @@ Lang.Blocks = {
     "behaviorConduct03013": "Traffic Accident",
     "behaviorConduct03014": "Elevator accident",
     "behaviorConduct03015": "Fine dust",
-    "behaviorConduct01001002": "Before the typhoon comes",
-    "behaviorConduct01001003": "When the typhoon watch occurs",
-    "behaviorConduct01001004": "When the typhoon warning occurs",
-    "behaviorConduct01001005": "After the typhoon has passed",
+    "behaviorConduct01001001": "Before the typhoon comes",
+    "behaviorConduct01001002": "When the typhoon warning occurs",
+    "behaviorConduct01001003": "After the typhoon has passed",
     "behaviorConduct01002002": "When the flood forecast occurs",
     "behaviorConduct01002003": "When the water is pushed",
     "behaviorConduct01002004": "After the water ir removed",
     "behaviorConduct01003002": "When the heavy rain forecast occurs",
     "behaviorConduct01003003": "When the heavy rain warning occurs",
     "behaviorConduct01003004": "After the heavy rain has passed",
-    "behaviorConduct01004002": "Before the gale comes",
-    "behaviorConduct01004003": "When the gale occurs",
-    "behaviorConduct01004004": "After the winds has passed",
-    "behaviorConduct01005002": "When the heavy snow occurs",
+    "behaviorConduct01004001": "Before the gale comes",
+    "behaviorConduct01004002": "When the gale occurs",
+    "behaviorConduct01005002": "Before the heavy snow comes",
+    "behaviorConduct01005003": "When the heavy snow occurs",
+    "behaviorConduct01005004": "After the heavy snow has passed",
+    "behaviorConduct01006001": "Before the cold wave comes",
     "behaviorConduct01006002": "When the cold wave occurs",
+    "behaviorConduct01007001": "When the storm warning occurs",
     "behaviorConduct01007002": "Before the storm comes",
-    "behaviorConduct01007003": "When the storm warning occurs",
-    "behaviorConduct01008002": "When the dust storm forecast occurs",
-    "behaviorConduct01008003": "When the dust storm occurs",
-    "behaviorConduct01008004": "After the dust storm has passed",
+    "behaviorConduct01008001": "When the dust storm forecast occurs",
+    "behaviorConduct01008002": "When the dust storm occurs",
+    "behaviorConduct01008003": "After the dust storm has passed",
     "behaviorConduct01009001": "Before the heatwave comes",
     "behaviorConduct01009002": "When the heatwave watch occurs",
-    "behaviorConduct01009003": "When the heatwave warning occurs",
-    "behaviorConduct01010001": "Before the drought comes",
+    "behaviorConduct01010001": "When the drought occurs",
     "behaviorConduct01011009": "Before the earthquakes comes",
     "behaviorConduct01011010": "When the earthquake occurs",
     "behaviorConduct01011012": "After the earthquake evacuation",
     "behaviorConduct01013002": "Before the tsunami comes",
     "behaviorConduct01013003": "When the tsunami occurs",
     "behaviorConduct01014001": "When the landslide watch occurs",
-    "behaviorConduct01014002": "When the landslide warning occurs",
     "behaviorConduct03002002": "treat burns",
     "behaviorConduct03002003": "prevate fever",
     "behaviorConduct03002004": "treat when bitten by a snake",
@@ -2193,7 +2192,9 @@ Lang.Buttons = {
     "multi_line": "Multiline",
     "lecture_finish_popup_reset_lecture": "Do it again",
     "lecture_finish_popup_show_output": "See the exercised",
-    "default_hint_title": "How do you make it?"
+    "lecture_view_hints": "View hints",
+    "default_hint_title": "How do you make it?",
+    "show_only_vector": "Only Vector"
 };
 Lang.Menus = {
     "nothing_to_import": "There is no content to upload.",
@@ -2550,13 +2551,18 @@ Lang.Menus = {
     "draw_new_ques_2": "do you want to move to draw?",
     "draw_new_go": "Move",
     "draw_new_stay": "No, Stay",
+    "file_upload_warn_title_image": "The following pictures are subject to the Terms of Use and related laws.",
+    "file_upload_warn_desc_image": "Violent and brutal picture<br>Sacred body exposure picture<br>Offensive or disgusting picture<br>Picture of copyright prohibited from unauthorized use",
+    "file_upload_warn_title_sound": "The following sounds are subject to the Terms of Use and related laws.",
+    "file_upload_warn_desc_sound": "Violent and cruel sound<br>Suggestive sound<br>Offensive or disgusting sound<br>Sound of copyright prohibited from unauthorized use",
+    "file_upload_warn_link": "Learn about copyright",
     "file_upload_desc_1": "The following pictures are subject to the Terms of Use and related laws.",
     "file_upload_desc_2": "Violent and brutal picture",
     "file_upload_desc_3": "Sacred body exposure picture",
     "file_upload_desc_4": "Offensive or disgusting picture",
     "file_upload_desc_5": "* Uploading of above contents may be restricted under user policy and related laws.",
-    "picture_upload_warn_1": "You can only upload jpg, png, bmp files that are less than 10MB.",
-    "sprite_upload_warn": "You can only upload image or object files such as jpg, png, bmp, eo that are less than 10MB.",
+    "picture_upload_warn_1": "You can only upload jpg, png, bmp, svg files that are less than 10MB.\nThe jpg, png, and bmp images are edited in bitmap mode and the svg image is in vector mode.",
+    "sprite_upload_warn": "You can only upload images in jpg, png, bmp, svg format or objects in eo format that are less than 10MB.",
     "sound_upload_warn_1": "You can only upload mp3 files that are less than 10MB.",
     "sound_edit_warn": "Please wait!<br>We're working on the Sound Editing function.",
     "file_upload_loading": "Loading",
@@ -2646,6 +2652,16 @@ Lang.Menus = {
     "lecture_finish_popup_title": "re-learn",
     "lecture_finish_popup_content_1": "Practice Completion Do you want to practice one step from the beginning?",
     "lecture_finish_popup_content_2": "* The practice completion status is maintained",
+    "lecture_start_title": "Start",
+    "lecture_goal_title": "Goal",
+    "lecture_goal_content": "Goal",
+    "lecture_prepare_title": "Prepare",
+    "lecture_summary_title": "Summary",
+    "lecture_summary_content": "Let's summarize what we learned.",
+    "lecture_quiz_title": "Quiz",
+    "lecture_quiz_content": "Let's check what we learned through the verification problem.",
+    "lecture_quiz_check_answer": "Check the answer",
+    "default_hint_title": "How do you make it?",
     "only_pdf": "Only pdf file is uploadable",
     "enter_project_video": "Enter at least one project or video",
     "enter_title": "Enter title.",
@@ -3439,7 +3455,7 @@ Lang.Menus = {
     "basic_mission_content": "A power outage at a toy robot factory magically enabled Entrybot to come to life. Help Entrybot escape the factory and find freedom!",
     "application_mission": "Application mission: Entry bot's Space Odyssey",
     "write_article": "Write",
-    "write_contest": "Apply",
+    "write_contest": "Submit",
     "view_all_articles": "View all posts",
     "view_own_articles": "View my posts",
     "view_qna_articles": "View no answered posts",
@@ -4229,6 +4245,8 @@ Lang.Menus = {
     "kumsungMid_edu_entry": "금성출판사 엔트리 학습하기",
     "gilbutEdu_edu_entry": "길벗출판사 엔트리 학습하기",
     "miraenCoding_edu_entry": "미래엔 엔트리 학습하기",
+    "textbook_help_data": "Help data",
+    "textbook_teaching_activity": "teaching and activity",
     "textbook_kyohacksa": "Kyohaksa",
     "textbook_kumsung": "Kumsung Publishing",
     "textbook_donga": "Dong-a Publishing",
@@ -4526,6 +4544,16 @@ Lang.Menus = {
     "group_add_course_3": "'Learn > Open Lessons > Courseware",
     "group_add_course_4": "Browse courseware",
     "hw_main_title": "Download Program",
+    "painter_contest_title": "Making a creative project",
+    "painter_content_head": "New Paint and New Objects",
+    "painter_content_body1": "How to make cool project with editing",
+    "painter_content_body2": "vector images and new objects",
+    "painter_content_button": "View Campaigns",
+    "campaign_contest_title": "Let's make a beautiful internet world",
+    "campaign_content_head": "Entry Project Contest",
+    "campaign_content_body1": "Create an originative project about 'internet ethics'",
+    "campaign_content_body2": "and apply for the contest!",
+    "campaign_content_button": "View details",
     "code_management_1": "Entry campaign #1 How to manage code",
     "code_management_2": "My Storage and Memo",
     "code_management_3": "Code management secret using my storage and memo!",
@@ -4678,6 +4706,7 @@ Lang.Menus = {
     "discuss_write_textarea_placeholer_tips": "This community is for Tips&Tricks.\nLet's make a fun community with consideration for each other!",
     "discuss_write_textarea_placeholer_free": "This community is for Discussion.\nLet's make a fun community with consideration for each other!",
     "discuss_write_textarea_placeholer_report": "This community is for Suggestions.\nLet's make a fun community with consideration for each other!",
+    "discuss_write_textarea_placeholer_contest2019": "※ 아래 양식으로 작성하여 제출해주세요.\n\n1. 학년 (예: 초등학교 5학년)\n\n2. 작품 URL (‘공유하기 > 작품 공유하기’에 작품을 공유한 다음, 작품 URL을 복사하여 적어주세요.)\n\n3. 작품 설명\n\n※ 공모전 수상자에게는 이메일이 발송될 예정입니다.\n엔트리의 ‘내 정보 수정‘에 이메일 정보를 꼭 입력해주세요!",
     "maze_road": "road",
     "account_deletion": "Account deletion",
     "bug_report_too_many_request": "Your report is being sent. Please try again later.",
@@ -4861,6 +4890,9 @@ Lang.Users = {
     "not_valid_email": "Please enter a correct type of email."
 };
 Lang.Workspace = {
+    "picture_rotation": "xoay vòng",
+    "picture_size": "kích cỡ",
+    "thickness": "độ dày",
     "follow": "Follow",
     "SaveWithPicture": "You have a unsaved drawing. Do you want to save it?",
     "RecursiveCallWarningTitle": "Limitation of function call",
@@ -4878,6 +4910,7 @@ Lang.Workspace = {
     "block_helper": "Khối giúp đỡ",
     "new_project": "New Project",
     "add_object": "thêm đối tượng",
+    "download_image_name": "entry block",
     "all": "Tất cả",
     "animal": "Animal",
     "arduino_entry": "Program for connect to Arduino",
@@ -5038,7 +5071,7 @@ Lang.Workspace = {
     "Variable_use_all_objects": "sử dụng cho tất cả các đối tượng",
     "Variable_use_this_object": "sử dụng cho đối tượng này",
     "Variable_used_at_all_objects": "biến số được sử dụng trên tất cả các đối tượng",
-    "Variable_create_cloud": "Cần biến số chia sẻ. <br>(Đã được lưu trên máy chủ)",
+    "Variable_create_cloud": "Cần biến số chia sẻ. \n(Đã được lưu trên máy chủ)",
     "Variable_used_at_special_object": "biến số được sử dụng trên đối tượng đặc biệt",
     "Variable_used_objects": "Object used for the variable",
     "draw_new": "Draw new",
@@ -5049,29 +5082,6 @@ Lang.Workspace = {
     "painter_edit": "biên soạn▼",
     "get_file": "nhận được",
     "copy_file": "sao chép",
-    "cut_picture": "cắt",
-    "paste_picture": "dán",
-    "remove_all": "xoá tất cả",
-    "new_picture": "hình ảnh mới",
-    "picture_size": "kích cỡ",
-    "picture_rotation": "xoay vòng",
-    "flip_horizontal": "flip horizontal",
-    "flip_vertical": "flip vertical",
-    "thickness": "độ dày",
-    "regular": "Regular",
-    "bold": "Bold",
-    "font_underline": "underline",
-    "font_tilt": "tilt",
-    "font_cancel": "cancel",
-    "font_color": "font color",
-    "font_fill": "Fill",
-    "font_family": "font",
-    "align_left": "left align ",
-    "align_center": "center align",
-    "align_right": "right align",
-    "italic": "Italic",
-    "textStyle": "chữ",
-    "add_picture": "Add Shape",
     "select_picture": "Select Library",
     "select_sound": "Select sound",
     "Size": "kích cỡ",
@@ -5101,7 +5111,7 @@ Lang.Workspace = {
     "robot_guide": "Robot Guide",
     "python_guide": "Entry Python Guide",
     "show_list_workspace": "hiển thị danh sách",
-    "List_create_cloud": "sử dụng cho danh sách chia sẻ <br> (được lưu trữ trong máy chủ)",
+    "List_create_cloud": "sử dụng cho danh sách chia sẻ \n(được lưu trữ trong máy chủ)",
     "confirm_quit": "You didn't saved your changes.",
     "confirm_load_temporary": "Bạn chưa lưu dự án. Bạn có muốn mở không?",
     "login_to_save": "Try to save after login.",
@@ -5178,7 +5188,9 @@ Lang.Workspace = {
     "brightness": "Brightness",
     "spoid": "Spoid",
     "palette_mode": "Palette mode",
-    "slider_mode": "Slider mode"
+    "slider_mode": "Slider mode",
+    "toast_error_title_object_center": "Caution",
+    "toast_error_contents_object_center": "If the center point is far from the object and goes out of the canvas, it returns to its original position."
 };
 Lang.code = "view code";
 Lang.EntryStatic = {
@@ -5301,6 +5313,7 @@ Lang.EntryStatic = {
     "general": "general",
     "curriculum_is_open_true": "open",
     "curriculum_open_false": "closed",
+    "contest2019": "Contest",
     "exhibit": "Contest",
     "notice": "Notice",
     "qna": "Q&A",
@@ -5894,14 +5907,15 @@ Lang.Helper = {
     "get_today_city_temperature": "<br>오늘 선택한 지역과 시간대의 기온을 의미합니다.<br> * 옵션<br> 시도: 서울, 강원, 경기, 경남, 경북, 광주, 대구, 대전, 부산, 세종, 울산, 인천, 전남, 전북, 제주, 충남, 충북 <br> 시군구: 시도별로 다름<br> 시간대: 00, 03, 06, 09, 12, 15, 18, 21",
     "count_festival": "<br>선택한 날짜와 지역에서 열리는 행사의 수입니다.<br> *옵션<br> 지역: 서울, 강원, 경기, 경남, 경북, 광주, 대구, 대전, 부산, 세종, 울산, 인천, 전남, 전북, 제주, 충남, 충북<br> 행사일: 1월~12월",
     "get_festival_info": "<br>선택한 날짜와 지역에서 열리는 행사에 대한 정보입니다.<br> *옵션<br> 지역: 서울, 강원, 경기, 경남, 경북, 광주, 대구, 대전, 부산, 세종, 울산, 인천, 전남, 전북, 제주, 충남, 충북<br> 행사일: 1월~12월<br> 정보: 제목, 지역, 시작일, 종료일, 주소, 개요, 홈페이지",
-    "get_translated_string": "<br>입력한 문자값을 선택한 언어로 번역합니다. 입력은 20자까지 가능합니다.<br> *번역 가능한 언어<br> - 한국어 → 영어, 중국어 간체, 중국어 번체, 베트남어, 스페인어, 인도네시아어, 태국어, 프랑스어<br> - 영어 → 일본어, 프랑스어, 한국어<br> - 중국어 간체, 중국어 번체, 베트남어, 스페인어, 인도네시아어, 태국어, 프랑스어 → 한국어<br> - 일본어 →영어<br> - 프랑스어 → 영어",
-    "check_language": "<br>입력된 문자값의 언어를 감지합니다. 입력은 문장 형태로 20자까지 가능합니다.",
-    "count_disaster_behavior": "<br>선택한 자연재해가 발생했을 때 해야할 행동요령의 수입니다.<br> *옵션<br> 자연재해: 태풍, 홍수, 호우, 강풍, 대설, 한파, 풍랑, 황사, 폭염, 가뭄, 지진, 해일, 산사태<br> 상황: 자연재해에 따라 다름 ",
-    "get_disaster_behavior": "<br>선택한 자연재해가 발생했을 때 해야하는 행동요령에 대한 정보입니다.<br> *옵션<br> 자연재해: 태풍, 홍수, 호우, 강풍, 대설, 한파, 풍랑, 황사, 폭염, 가뭄, 지진, 해일, 산사태<br> 상황: 자연재해에 따라 다름",
-    "count_lifeSafety_behavior": "<br>선택한 생활 속 문제가 발생했을 때 해야할 행동요령의 수입니다.<br> *옵션<br> 생활 속 문제: 응급처치, 심폐소생술, 소화기·소화전 사용법, 식중독, 산행안전, 실종유괴, 성폭력, 학교폭력, 가정 폭력, 억류 및 납치, 교통사고, 승강기 안전사고, 미세먼지<br> 상황: 생활 속 문제에 따라 다름 ",
-    "get_lifeSafety_behavior": "<br>선택한 생활 속 문제가 발생했을 때 해야할 행동요령에 대한 정보입니다.<br> *옵션<br> 생활 속 문제: 응급처치, 심폐소생술, 소화기·소화전 사용법, 식중독, 산행안전, 실종유괴, 성폭력, 학교폭력, 가정 폭력, 억류 및 납치, 교통사고, 승강기 안전사고, 미세먼지<br> 상황: 생활 속 문제에 따라 다름 ",
-    "read_text": "입력한 문자값을 설정된 목소리로 읽습니다.<br>입력은 2500자까지 가능합니다.<br>인터넷에 연결되어 있지 않거나 인터넷 환경이 불안할 경우, 해당 블록이 실행되지 않고 다음 블록으로 넘어갈 수 있습니다.",
-    "set_tts_property": "선택한 목소리가 선택한 속도와 선택한 음높이로 설정됩니다."
+    "get_translated_string": "The number of events held on the selected date and region.",
+    "check_language": "Reports the information(Title, area, start date, end date, address, overview, homepage) about events held on the selected date and region.",
+    "count_disaster_behavior": "<br>The number of behavior-guideline's when selected natural disaster occurs.",
+    "get_disaster_behavior": "<br>Reports the information about behavior-guidelines when selected natural disaster occurs.",
+    "count_lifeSafety_behavior": "<br>The number of behavior-guideline's when selected life problem occurs.",
+    "get_lifeSafety_behavior": "<br>Reports the information about behavior-guidelines when selected life problem occurs.",
+    "read_text": "Read the entered text in setted voice.",
+    "set_tts_property": "Set the selected voice to selected speed and selected pitch. <br> The selected voice will be set up to selected speed and selected pitch.",
+    "read_text_wait_with_block": "Read the entered text in setted voice then the next block runs after the reading finished."
 };
 Lang.Category = {
     "entrybot_friends": "Entrybot friends",
@@ -5940,6 +5954,9 @@ Lang.Category = {
     "stuff_others": "Others",
     "fantasy": "Fantasy",
     "interface": "Interface",
+    "interface_website": "Website",
+    "interface_game": "Game",
+    "interface_others": "Others",
     "background": "Background",
     "background_outdoor": "Outdoor",
     "background_indoor": "Indoor",
@@ -6033,7 +6050,14 @@ Lang.Fonts = {
     "yisunshin": "YiSunShin",
     "goyang": "Goyang",
     "flower_road": "SangSangFlowerRoad",
-    "uhbeemysen": "UhBeemysen"
+    "uhbeemysen": "UhBeemysen",
+    "sd_comic_stencil": "Sandoll Comicstencil",
+    "sd_childfundkorea": "Sandoll Childfundkorea",
+    "sd_cinema_theater": "Sandoll Cinema Theater",
+    "sd_mapssi": "Sandoll Mapssi",
+    "sd_shabang": "Sandoll Shabang",
+    "sd_woodcarving": "Sandoll Woodcarving",
+    "sd_yongbi": "Sandoll Yongbi"
 };
 Lang.Hw = {
     "note": "Nốt nhạc",
@@ -6551,6 +6575,7 @@ Lang.template = {
     "tts_title": "%1",
     "read_text": "read %1 %2",
     "set_tts_property": "set in a %1 voice at %2 speed and %3 pitch %4",
+    "read_text_wait_with_block": "read %1 and wait %2",
     "load_expansion_block": "add",
     "expansion_block_descriptions": "The expansion block works normally when the Internet is connected.",
     "weather_title_text": "weather",
@@ -7588,7 +7613,7 @@ Lang.PythonHelper = {
     "quotient_and_mod_exampleCode": "def when_start():\n    Entry.print_for_sec(10 // 3, 2)\n    Entry.print_for_sec(10 % 3, 2)",
     "quotient_and_mod_exampleDesc": "[시작하기]버튼을 클릭하면 10 / 3의 몫인 3을 2초 동안 말하고, 나머지인 1을 2초 동안 말합니다.",
     "calc_operation_desc": "A의 연산값입니다.",
-    "calc_operation_elements": "A, B-- 연산하고자 하는 숫자값<br>① A ** 2 : A를 제곱한 값<br>② math.sqrt(A): A의 루트값<br>③ math.sin(A): A의 사인값<br>④ math.cos(A): A의 코사인 값<br>⑤ math.tan(A):  A의 탄젠트값 <br>⑥ math.asin(A): A의 아크사인값<br>⑦ math.acos(A): A의 아크코사인값<br>⑧ math.atan(): A의 아크탄젠트값<br>⑨ math.log10(A): A의 로그값<br>⑩ math.log(A): A의 자연로그값<br>⑪ A - math.floor(A): A의 소수점 부분<br>⑫ math.floor(A): A의 소수점 버림값<br>⑬ math.ceil(A): A의 소수점 올림값<br>⑭ math.round(A): A의 반올림값<br>⑮ math.factorial(A): A의 팩토리얼 값<br>⑯ math.fabs(A): A의 절댓값",
+    "calc_operation_elements": "A, B-- 연산하고자 하는 숫자값<br>① A ** 2 : A를 제곱한 값<br>② math.sqrt(A): A의 루트값<br>③ math.sin(A): A의 사인값<br>④ math.cos(A): A의 코사인 값<br>⑤ math.tan(A):  A의 탄젠트값 <br>⑥ math.asin(A): A의 아크사인값<br>⑦ math.acos(A): A의 아크코사인값<br>⑧ math.atan(): A의 아크탄젠트값<br>⑨ math.log10(A): A의 로그값<br>⑩ math.log(A): A의 자연로그값<br>⑪ A - math.floor(A): A의 소수점 부분<br>⑫ math.floor(A): A의 소수점 버림값<br>⑬ math.ceil(A): A의 소수점 올림값<br>⑭ math.round(A): A의 소수점 반올림값<br>⑮ math.factorial(A): A의 팩토리얼 값<br>⑯ math.fabs(A): A의 절댓값",
     "calc_operation_exampleCode": "def when_start():\n    Entry.print_for_sec(10 ** 2, 2)\n    Entry.print_for_sec(math.sqrt(9), 2)\n    Entry.print_for_sec(math.sin(90), 2)\n    Entry.print_for_sec(math.fabs(-10), 2)",
     "calc_operation_exampleDesc": "[시작하기]버튼을 클릭하면 10의 제곱, 9의 루트값, 90의 사인값, -10의 절댓값을 각 2초 동안 말합니다.",
     "get_project_timer_value_desc": "이 명령이 실행되는 순간 초시계에 저장된 값입니다.",
@@ -7760,15 +7785,54 @@ Lang.PythonHelper = {
 Lang.Painter = {
     "select_cut": "You can cut with a square dotted line.",
     "thickness_of_eraser": "Thickness of eraser",
-    "font": "font",
+    "font": "Font",
     "font_color": "Font color",
     "outline_color": "Outline color",
+    "outline_thickness": "Outline thickness",
     "fill_color": "Fill",
     "inversion": "Inversion",
     "font_size": "Font size",
     "font_family": "Font family",
     "fullscreen": "Full screen",
-    "exit_fullscreen": "Exit full screen"
+    "exit_fullscreen": "Exit full screen",
+    "picture_import": "Import a picture",
+    "edit_way": "Edit way",
+    "line": "Hàng",
+    "vertex": "Vertex",
+    "bitmap": "Bitmap",
+    "vector": "Vector",
+    "forward": "đằng trước",
+    "backward": "đằng sau",
+    "front": "phía trước",
+    "back": "phía sau",
+    "group": "group",
+    "ungroup": "ungroup",
+    "copy_picture": "sao chép",
+    "delete_picture": "delete",
+    "cut_picture": "cắt",
+    "paste_picture": "dán",
+    "remove_all": "xoá tất cả",
+    "new_picture": "hình ảnh mới",
+    "edit": "Edit",
+    "picture_size": "kích cỡ",
+    "picture_rotation": "xoay vòng",
+    "thickness": "độ dày",
+    "font_fill": "Fill",
+    "painter_file_save": "lưu",
+    "painter_file_saveas": "lưu thành hình ảnh mới",
+    "file": "File",
+    "Pencil": "Cây bút",
+    "Line": "Hàng",
+    "Path": "Path",
+    "Pan": "Move",
+    "Select": "Select",
+    "Rectangle": "Hình chữ nhật",
+    "Ellipse": "hình bầu dục",
+    "Text": "Hộp văn bản",
+    "Fill": "Lấp đầy",
+    "Eraser": "Cục tẩy",
+    "get_file": "nhận được",
+    "copy_file": "sao chép"
 };
 
 Lang.Support = ["ko", "en", "vn", "jp"];
