@@ -599,7 +599,7 @@ EntryStatic.objectSubCategories = {
     environment: ['environment_nature', 'environment_space', 'environment_others'],
     stuff: ['stuff_living', 'stuff_hobby', 'stuff_others'],
     fantasy: [],
-    interface: [],
+    interface: ['interface_website', 'interface_game', 'interface_others'],
     background: [
         'background_outdoor',
         'background_indoor',
@@ -676,6 +676,55 @@ Object.defineProperty(EntryStatic, 'fonts', {
                 name: Lang.Fonts.uhbeemysen,
                 family: 'UhBeemysen',
                 url: '/css/uhbeemysen.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_comic_stencil,
+                family: 'SDComicStencil',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDComicStencil.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_childfundkorea,
+                family: 'SDChildfundkorea',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDChildfundkorea.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_cinema_theater,
+                family: 'SDCinemaTheater',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDCinemaTheater.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_mapssi,
+                family: 'SDMapssi',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDMapssi.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_shabang,
+                family: 'SDShabang',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDShabang.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_woodcarving,
+                family: 'SDWoodcarving',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDWoodcarving.css',
+                visible: true,
+            },
+            {
+                name: Lang.Fonts.sd_yongbi,
+                family: 'SDYongbi',
+                style: { backgroundColor: '#f7fcff', padding: '17px 22px 14px 16px', margin:'0px', borderTop: 'solid 1px #d6e9f4' },
+                url: '/css/SDYongbi.css',
                 visible: true,
             },
         ];
