@@ -355,13 +355,13 @@ EntryStatic.getAllBlocks = function() {
         {
             category: 'hw_robot',
             blocks: [
-                'arduino_download_connector',
-                'download_guide',
-                'arduino_connected',
-                'arduino_connect',
                 'robot_reconnect',
                 'arduino_open',
                 'arduino_cloud_pc_open',
+                'arduino_connect',
+                'arduino_download_connector',
+                'download_guide',
+                'arduino_connected',
             ],
         },
         {
@@ -551,7 +551,7 @@ EntryStatic.objectSubCategories = {
     environment: ['environment_nature', 'environment_space', 'environment_others'],
     stuff: ['stuff_living', 'stuff_hobby', 'stuff_others'],
     fantasy: [],
-    interface: [],
+    interface: ['interface_website', 'interface_game', 'interface_others'],
     background: [
         'background_outdoor',
         'background_indoor',
@@ -625,6 +625,48 @@ EntryStatic.fonts = [
         name: Lang.Fonts.uhbeemysen,
         family: 'UhBeemysen',
         url: '/css/uhbeemysen.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_comic_stencil,
+        family: 'SDComicStencil',
+        url: '/css/SDComicStencil.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_childfundkorea,
+        family: 'SDChildfundkorea',
+        url: '/css/SDChildfundkorea.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_cinema_theater,
+        family: 'SDCinemaTheater',
+        url: '/css/SDCinemaTheater.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_mapssi,
+        family: 'SDMapssi',
+        url: '/css/SDMapssi.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_shabang,
+        family: 'SDShabang',
+        url: '/css/SDShabang.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_woodcarving,
+        family: 'SDWoodcarving',
+        url: '/css/SDWoodcarving.css',
+        visible: true,
+    },
+    {
+        name: Lang.Fonts.sd_yongbi,
+        family: 'SDYongbi',
+        url: '/css/SDYongbi.css',
         visible: true,
     },
 ];
