@@ -3,17 +3,17 @@
  */
 'use strict';
 
-
-
 /**
  * Helper provide block description with 'blockHelper'
  */
 Entry.Helper = function() {
     this.visible = false;
-    Entry.addEventListener('workspaceChangeMode', function() {
-        if (this._blockView)
-            this.renderBlock(this._blockView.type)
-    }.bind(this));
+    Entry.addEventListener(
+        'workspaceChangeMode',
+        function() {
+            if (this._blockView) this.renderBlock(this._blockView.type);
+        }.bind(this)
+    );
     this.resize = Entry.Utils.debounce(this.resize, 300);
 };
 
@@ -25,43 +25,36 @@ p.generateView = function(parentView, option) {
     this.parentView_ = parentView;
     var helper = this;
     helper.blockHelpData = EntryStatic.blockInfo;
-    var blockHelperWrapper = Entry.createElement('div',
-        'entryBlockHelperWorkspaceWrapper');
-    var blockHelperView = Entry.createElement('div',
-                            'entryBlockHelperWorkspace');
+    var blockHelperWrapper = Entry.createElement('div', 'entryBlockHelperWorkspaceWrapper');
+    var blockHelperView = Entry.createElement('div', 'entryBlockHelperWorkspace');
     blockHelperWrapper.appendChild(blockHelperView);
     this.view = blockHelperWrapper;
-    if (Entry.isForLecture)
-        blockHelperView.addClass('lecture');
+    if (Entry.isForLecture) blockHelperView.addClass('lecture');
     helper.parentView_.appendChild(blockHelperWrapper);
 
-    var blockHelperContent =
-        Entry.createElement('div', 'entryBlockHelperContentWorkspace');
+    var blockHelperContent = Entry.createElement('div', 'entryBlockHelperContentWorkspace');
     this._contentView = blockHelperContent;
 
     var commandTitle = Entry.createElement('div');
     commandTitle.addClass('entryBlockHelperTitle textModeElem');
-    commandTitle.innerHTML = '명령어';
+    commandTitle.innerHTML = 'Command';
     blockHelperContent.appendChild(commandTitle);
 
     blockHelperContent.addClass('entryBlockHelperIntro');
-    if (Entry.isForLecture)
-        blockHelperContent.addClass('lecture');
+    if (Entry.isForLecture) blockHelperContent.addClass('lecture');
     blockHelperView.appendChild(blockHelperContent);
     helper.blockHelperContent_ = blockHelperContent;
     helper.blockHelperView_ = blockHelperView;
 
-    var blockHelperBlock = Entry.createElement('div',
-                            'entryBlockHelperBlockWorkspace');
+    var blockHelperBlock = Entry.createElement('div', 'entryBlockHelperBlockWorkspace');
     helper.blockHelperContent_.appendChild(blockHelperBlock);
 
     var descTitle = Entry.createElement('div');
     descTitle.addClass('entryBlockHelperTitle textModeElem');
-    descTitle.innerHTML = '설명';
+    descTitle.innerHTML = 'Explanation';
     blockHelperContent.appendChild(descTitle);
 
-    var blockHelperDescription = Entry.createElement('div',
-                            'entryBlockHelperDescriptionWorkspace');
+    var blockHelperDescription = Entry.createElement('div', 'entryBlockHelperDescriptionWorkspace');
     blockHelperDescription.addClass('entryBlockHelperContent selectAble');
     helper.blockHelperContent_.appendChild(blockHelperDescription);
     blockHelperDescription.innerHTML = Lang.Helper.Block_click_msg;
@@ -69,13 +62,11 @@ p.generateView = function(parentView, option) {
 
     var elementsTitle = Entry.createElement('div');
     elementsTitle.addClass('entryBlockHelperTitle textModeElem');
-    elementsTitle.innerHTML = '요소';
+    elementsTitle.innerHTML = 'Element';
     blockHelperContent.appendChild(elementsTitle);
     this._elementsTitle = elementsTitle;
 
-
-    this._elementsContainer =
-        Entry.createElement('div', 'entryBlockHelperElementsContainer');
+    this._elementsContainer = Entry.createElement('div', 'entryBlockHelperElementsContainer');
 
     this._elementsContainer.addClass('entryBlockHelperContent textModeElem selectAble');
     blockHelperContent.appendChild(this._elementsContainer);
@@ -83,7 +74,7 @@ p.generateView = function(parentView, option) {
     if (typeof CodeMirror !== 'undefined') {
         var codeMirrorTitle = Entry.createElement('div');
         codeMirrorTitle.addClass('entryBlockHelperTitle textModeElem');
-        codeMirrorTitle.innerHTML = '예시 코드';
+        codeMirrorTitle.innerHTML = 'Example Code';
         blockHelperContent.appendChild(codeMirrorTitle);
 
         var codeMirrorView = Entry.createElement('div', 'entryBlockHelperCodeMirrorContainer');
@@ -92,13 +83,13 @@ p.generateView = function(parentView, option) {
 
         this.codeMirror = CodeMirror(codeMirrorView, {
             lineNumbers: true,
-            value: "",
-            mode: {name: "python"},
+            value: '',
+            mode: { name: 'python' },
             indentUnit: 4,
-            theme: "default",
+            theme: 'default',
             viewportMargin: 10,
             styleActiveLine: false,
-            readOnly: true
+            readOnly: true,
         });
 
         this._doc = this.codeMirror.getDoc();
@@ -106,7 +97,7 @@ p.generateView = function(parentView, option) {
 
         var codeMirrorDescTitle = Entry.createElement('div');
         codeMirrorDescTitle.addClass('entryBlockHelperTitle textModeElem');
-        codeMirrorDescTitle.innerHTML = '예시 설명';
+        codeMirrorDescTitle.innerHTML = 'Example Explanation';
         blockHelperContent.appendChild(codeMirrorDescTitle);
 
         this._codeMirrorDesc = Entry.createElement('div');
@@ -125,14 +116,13 @@ p.generateView = function(parentView, option) {
 p.bindWorkspace = function(workspace) {
     if (!workspace) return;
 
-    if (this._blockViewObserver)
-        this._blockViewObserver.destroy();
+    if (this._blockViewObserver) this._blockViewObserver.destroy();
 
     this.workspace = workspace;
-    if (this._renderView)
-        this._renderView.workspace = workspace;
-    this._blockViewObserver =
-        workspace.observe(this, "_updateSelectedBlock", ['selectedBlockView']);
+    if (this._renderView) this._renderView.workspace = workspace;
+    this._blockViewObserver = workspace.observe(this, '_updateSelectedBlock', [
+        'selectedBlockView',
+    ]);
 };
 
 /**
@@ -149,8 +139,7 @@ p._updateSelectedBlock = function() {
 
 p.renderBlock = function(type) {
     var description = Lang.Helper[type];
-    if (!type || !this.visible || !description || Entry.block[type].isPrimitive)
-        return;
+    if (!type || !this.visible || !description || Entry.block[type].isPrimitive) return;
 
     if (this.first) {
         this.blockHelperContent_.removeClass('entryBlockHelperIntro');
@@ -159,7 +148,7 @@ p.renderBlock = function(type) {
 
     var code = this.code;
     code.clear();
-    var def = Entry.block[type].def || {type: type};
+    var def = Entry.block[type].def || { type: type };
 
     if (this.workspace.getMode() === Entry.Workspace.MODE_VIMBOARD) {
         this._contentView.addClass('textMode');
@@ -186,12 +175,12 @@ p.renderBlock = function(type) {
                     box.appendChild(left);
                     box.appendChild(right);
                     this._elementsContainer.appendChild(box);
-                }.bind(this))(elements.shift());
+                }.bind(this)(elements.shift()));
             }
         } else {
             this._elementsTitle.addClass('entryRemove');
         }
-        this._codeMirrorDesc.innerHTML = (Lang.PythonHelper[type + '_exampleDesc']);
+        this._codeMirrorDesc.innerHTML = Lang.PythonHelper[type + '_exampleDesc'];
 
         var exampleCode = Lang.PythonHelper[type + '_exampleCode'];
         this._codeMirror.setValue(exampleCode);
