@@ -1,105 +1,111 @@
-'use strict';
+import Extension from '../extensions/extension';
 
-Entry.Painter = function(view) {
-    this.view = view;
-    this.baseUrl = Entry.painterBaseUrl;
+let EntryPaint;
+Entry.Painter = class Painter {
+    constructor(view) {
+        this.view = view;
+        this.cache = [];
 
-    this.file = {
-        id: Entry.generateHash(),
-        name: '새그림',
-        modified: false,
-        mode: 'new', // new or edit
-    };
+        this.file = {
+            id: Entry.generateHash(),
+            name: '새그림',
+            modified: false,
+            mode: 'new', // new or edit
+        };
 
-    this._keyboardEvents = [];
+        this.dropper = Extension.getExtension('Dropper');
+        this.isShow = false;
+        this.clipboard = null;
+        this._keyboardEvents = [];
+        Entry.addEventListener('pictureImport', this.addPicture.bind(this));
+        Entry.addEventListener('run', this.detachKeyboardEvents.bind(this));
+        Entry.addEventListener('stop', this.attachKeyboardEvents.bind(this));
+        this.importEntryPaint();
+    }
 
-    Entry.addEventListener('pictureImport', this.addPicture.bind(this));
-    Entry.addEventListener('run', this.detachKeyboardEvents.bind(this));
-    Entry.addEventListener('stop', this.attachKeyboardEvents.bind(this));
+    async importEntryPaint() {
+        EntryPaint = window.EntryPaint.default;
+        if (this.requestShow) {
+            this.initialize();
+        }
+    }
 
-    //dropdown header dispose
-    $('body').on('mouseup', '.active li', function() {
-        $('.painterTopHeader.active').removeClass('active');
-    });
+    get graphicsMode() {
+        return this.entryPaint.graphicsMode;
+    }
 
-    this.clipboard = null;
-};
-
-(function(p) {
-    p.initialize = function() {
-        if (this.lc) {
+    initialize() {
+        if (this.entryPaint || !EntryPaint) {
+            this.requestShow = true;
             return;
         }
 
-        const that = this;
+        this.isShow = true;
 
-        const imgURL = that.baseUrl;
-        const bgImage = new Image();
-        bgImage.src = `${imgURL}/transparent-pattern.png`;
+        this.entryPaint = EntryPaint.create({ parent: this.view, mode: 'entry' });
+        this.entryPaint.setDropper && this.entryPaint.setDropper(this.dropper);
 
-        const WIDTH = 960;
-        const HEIGHT = 540;
-        that.lc = LC.init(that.view, {
-            imageURLPrefix: imgURL,
-            zoomMax: 3.0,
-            zoomMin: 0.5,
-            toolbarPosition: 'bottom',
-            imageSize: { width: WIDTH, height: HEIGHT },
-            backgroundShapes: [
-                LC.createShape('Rectangle', {
-                    x: 0,
-                    y: 0,
-                    width: WIDTH,
-                    height: HEIGHT,
-                    strokeWidth: 0,
-                    strokeColor: 'transparent',
-                }),
-            ],
+        this.isImport = true;
+        this.entryPaint.on('SNAPSHOT_SAVED', (e) => {
+            if (!this.isImport && Entry.stage.selectedObject) {
+                Entry.do('editPicture', e, this.entryPaint);
+                this.file.modified = true;
+            }
+            this.isImport = false;
         });
-        //that.lc.respondToSizeChange();
+        this.entryPaint.on('NEW_PICTURE', this.newPicture.bind(this));
+        this.entryPaint.on('IMPORT_IMAGE', () => {
+            Entry.dispatchEvent('openPictureImport');
+        });
+        this.entryPaint.on('SAVE_PICTURE', () => {
+            this.fileSave(false);
+        });
+        this.entryPaint.on('SAVE_NEW_PICTURE', () => {
+            this.file.mode = 'new';
+            this.fileSave(false);
+        });
+        this.entryPaint.on('FULL_SCREEN_ON', () => {
+            this.toggleFullscreen(true);
+        });
+        this.entryPaint.on('FULL_SCREEN_OFF', () => {
+            this.toggleFullscreen(false);
+        });
 
-        bgImage.onload = function() {
-            that.lc.repaintLayer('background');
-        };
+        Entry.addEventListener('pictureSelected', this.changePicture.bind(this));
+    }
 
-        const watchFunc = function(e) {
-            if (e && ((e.shape && !e.opts && e.shape.isPass) || (e.opts && e.opts.isPass))) {
-                Entry.do('processPicture', e, that.lc);
-            } else {
-                Entry.do('editPicture', e, that.lc);
-            }
-            if (Entry.stage.selectedObject) {
-                that.file.modified = true;
-            }
-        };
-
-        that.lc.on('clear', watchFunc);
-        that.lc.on('remove', watchFunc);
-        that.lc.on('shapeEdit', watchFunc);
-        that.lc.on('shapeSave', watchFunc);
-
-        that.lc.on('toolChange', that.updateEditMenu.bind(that));
-        that.lc.on('lc-pointerdrag', that.stagemousemove.bind(that));
-        that.lc.on('lc-pointermove', that.stagemousemove.bind(that));
-
-        that.initTopBar();
-        that.updateEditMenu();
-
-        that.attachKeyboardEvents();
-    };
-
-    p.show = function() {
-        if (!this.lc) {
+    show() {
+        if (!this.isShow) {
             this.initialize();
         }
-        this.isShow = true;
-    };
+        this.realign && Entry.windowResized.detach(this.realign);
+        this.realign = Entry.windowResized.attach(this.view, this.entryPaint.realign);
+    }
 
-    p.hide = function() {
-        this.isShow = false;
-    };
+    hide() {
+        this.alertSaveModifiedPicture();
+        this.entryPaint && Entry.windowResized.detach(this.realign);
+    }
 
-    p.changePicture = function(picture = {}) {
+    newPicture() {
+        const newPicture = {
+            dimension: {
+                height: 1,
+                width: 1,
+            },
+            fileurl: `${Entry.mediaFilePath}_1x1.png`,
+            name: Lang.Painter.new_picture,
+            imageType: 'png',
+        };
+
+        newPicture.id = Entry.generateHash();
+        if (this.file && this.file.objectId) {
+            newPicture.objectId = this.file.objectId;
+        }
+        Entry.playground.addPicture(newPicture, true);
+    }
+
+    changePicture(picture = {}, removed) {
         if (this.file && this.file.id === picture.id) {
             if (!this.file.isUpdate) {
                 Entry.stage.updateObject();
@@ -119,31 +125,38 @@ Entry.Painter = function(view) {
                 Entry.engine.toggleStop();
                 wasRun = true;
             }
-            entrylms.confirm(Lang.Menus.save_modified_shape).then(
-                function(result) {
-                    this.isConfirm = false;
-                    if (result === true) {
-                        this.file_save(true);
-                    } else {
-                        this.file.modified = false;
-                    }
 
-                    if (!wasRun) {
-                        this.afterModified(picture);
-                    } else {
-                        Entry.playground.injectPicture();
-                    }
-                }.bind(this)
-            );
+            if (removed) {
+                this.updatePicture(picture);
+            } else {
+                this.alertSaveModifiedPicture(picture, wasRun);
+            }
         }
         Entry.stage.updateObject();
         this.file.isUpdate = true;
-    };
+    }
 
-    p.afterModified = function(picture) {
+    updatePicture(picture = {}, wasRun = true, result = true) {
+        this.isConfirm = false;
+        result ? this.fileSave(true) : (this.file.modified = false);
+        wasRun ? Entry.playground.injectPicture() : this.afterModified(picture);
+        Entry.stage.updateObject();
+    }
+
+    alertSaveModifiedPicture(picture, wasRun) {
+        if (!this.file.modified) {
+            return;
+        }
+
+        entrylms.confirm(Lang.Menus.save_modified_shape).then((result) => {
+            this.updatePicture(picture, wasRun, result);
+        });
+    }
+
+    afterModified(picture) {
         const file = this.file;
         file.modified = false;
-        this.lc.clear(false);
+        this.isImport = true;
 
         if (picture.id) {
             file.id = picture.id || Entry.generateHash();
@@ -154,319 +167,75 @@ Entry.Painter = function(view) {
             this.addPicture(picture, true);
         } else {
             file.id = Entry.generateHash();
+            this.entryPaint.reset();
         }
-        // INFO: picture 변경시마다 undoStack 리셋
-        this.lc.undoStack = [];
+
         Entry.stateManager.removeAllPictureCommand();
-    };
+    }
 
-    p.addPicture = function(picture, isOriginal) {
-        const image = new Image();
+    getImageSrc(picture) {
+        const { fileurl } = picture || {};
+        if (fileurl) {
+            return fileurl;
+        }
 
-        if (picture.fileurl) {
-            image.src = picture.fileurl;
+        const { imageType = 'png', filename } = picture || {};
+        return `${Entry.defaultPath}/uploads/${filename.substring(0, 2)}/${filename.substring(
+            2,
+            4
+        )}/image/${filename}.${imageType}`;
+    }
+
+    addPicture(picture, isChangeShape) {
+        const { imageType = 'png' } = picture || {};
+        const imageSrc = this.getImageSrc(picture);
+
+        isChangeShape && (this.isImport = true);
+
+        switch (imageType) {
+            case 'png':
+                this.entryPaint.addBitmap(imageSrc, {
+                    graphicsMode: this.isImport ? this.graphicsMode.BITMAP : '',
+                });
+                break;
+            case 'svg':
+                this.entryPaint.addSVG(imageSrc, {
+                    graphicsMode: this.isImport ? this.graphicsMode.VECTOR : '',
+                });
+                break;
+        }
+    }
+
+    _getImageType() {
+        if (this.entryPaint.mode === this.graphicsMode.VECTOR) {
+            return 'svg';
         } else {
-            // deprecated
-            image.src = `${Entry.defaultPath}/uploads/${picture.filename.substring(
-                0,
-                2
-            )}/${picture.filename.substring(2, 4)}/image/${picture.filename}.png`;
+            return 'png';
         }
+    }
 
-        const dimension = picture.dimension;
-        const shape = LC.createShape('Image', {
-            x: 480,
-            y: 270,
-            width: dimension.width,
-            height: dimension.height,
-            image,
-        });
-
-        this.lc.saveShape(shape, !isOriginal);
-
-        image.onload = function() {
-            this.lc.setTool(this.lc.tools.SelectShape);
-            this.lc.tool.setShape(this.lc, shape);
-        }.bind(this);
-    };
-
-    p.copy = function() {
-        if (this.lc.tool.name !== 'SelectShape' || !this.lc.tool.selectedShape) {
-            return;
-        }
-
-        const shape = this.lc.tool.selectedShape;
-        this.clipboard = {
-            className: shape.className,
-            data: shape.toJSON(),
-        };
-        this.updateEditMenu();
-    };
-
-    p.cut = function() {
-        if (this.lc.tool.name !== 'SelectShape' || !this.lc.tool.selectedShape) {
-            return;
-        }
-
-        this.copy();
-        const shape = this.lc.tool.selectedShape;
-        this.lc.removeShape(shape);
-        this.lc.tool.setShape(this.lc, null);
-    };
-
-    p.paste = function() {
-        if (!this.clipboard) {
-            return;
-        }
-
-        const shape = this.lc.addShape(this.clipboard);
-        this.lc.setTool(this.lc.tools.SelectShape);
-        this.lc.tool.setShape(this.lc, shape);
-    };
-
-    p.updateEditMenu = function() {
-        const isSelected = this.lc.tool.name === 'SelectShape' ? 'block' : 'none';
-        // this._cutButton.style.display = isSelected;
-        // this._copyButton.style.display = isSelected;
-        // this._pasteButton.style.display = this.clipboard ? "block" : "none";
-    };
-
-    p.file_save = function(taskParam) {
+    fileSave(taskParam) {
         if (!Entry.stage.selectedObject) {
             return;
         }
-        this.lc.trigger('dispose');
-        const dataURL = this.lc.getImage().toDataURL();
-        this.file_ = JSON.parse(JSON.stringify(this.file));
+        const dataURL = this.entryPaint.getDataURL();
+        if (this.entryPaint.mode === this.graphicsMode.VECTOR) {
+            this.file.svg = this.entryPaint.exportSVG();
+        } else {
+            delete this.file.svg;
+        }
+        this.file.ext = this._getImageType();
+        const file = JSON.parse(JSON.stringify(this.file));
         Entry.dispatchEvent('saveCanvasImage', {
-            file: this.file_,
+            file,
             image: dataURL,
             task: taskParam,
         });
-
         this.file.isUpdate = false;
         this.file.modified = false;
-    };
+    }
 
-    p.newPicture = function() {
-        if (!Entry.stage.selectedObject) {
-            return;
-        }
-        const newPicture = {
-            dimension: {
-                height: 1,
-                width: 1,
-            },
-            //filename: "_1x1",
-            fileurl: `${Entry.mediaFilePath}_1x1.png`,
-            name: Lang.Workspace.new_picture,
-        };
-
-        newPicture.id = Entry.generateHash();
-        if (this.file && this.file.objectId) {
-            newPicture.objectId = this.file.objectId;
-        }
-        Entry.playground.addPicture(newPicture, true);
-    };
-
-    p._keyboardPressControl = function(e) {
-        if (!this.isShow || Entry.Utils.isInInput(e)) {
-            return;
-        }
-
-        const keyCode = e.keyCode || e.which;
-        const ctrlKey = e.ctrlKey;
-
-        if (keyCode == 8 || keyCode == 46) {
-            //destroy
-            this.cut();
-            e.preventDefault();
-        } else if (ctrlKey) {
-            if (keyCode == 67) {
-                //copy
-                this.copy();
-            } else if (keyCode == 88) {
-                //cut
-                this.cut();
-            }
-        }
-
-        if (ctrlKey && keyCode == 86) {
-            //paste
-            this.paste();
-        }
-        this.lc.trigger('keyDown', e);
-    };
-
-    p._keyboardUpControl = function(e) {
-        if (!this.isShow || Entry.Utils.isInInput(e)) {
-            return;
-        }
-
-        this.lc.trigger('keyUp', e);
-    };
-
-    p.toggleFullscreen = function(isFullscreen) {
-        const { painter = {}, pictureView_ } = Entry.playground;
-        const { view = {} } = painter;
-        const $view = $(view);
-        if ((isFullscreen !== true && $view.hasClass('fullscreen')) || isFullscreen === false) {
-            pictureView_.appendChild(view);
-            $(view).removeClass('fullscreen');
-            if (this.fullscreenButton) {
-                this.fullscreenButton.setAttribute('title', Lang.Painter.fullscreen);
-                this.fullscreenButton.setAttribute('alt', Lang.Painter.fullscreen);
-            }
-        } else {
-            document.body.appendChild(view);
-            $(view).addClass('fullscreen');
-            if (this.fullscreenButton) {
-                this.fullscreenButton.setAttribute('title', Lang.Painter.exit_fullscreen);
-                this.fullscreenButton.setAttribute('alt', Lang.Painter.exit_fullscreen);
-            }
-        }
-        $(view)
-            .find('.lc-drawing.with-gui')
-            .trigger('resize');
-    };
-
-    p.initTopBar = function() {
-        const painter = this;
-
-        const ce = Entry.createElement;
-
-        const painterTop = ce(document.getElementById('canvas-top-menu'));
-        painterTop.addClass('entryPlaygroundPainterTop');
-        painterTop.addClass('entryPainterTop');
-
-        const painterTopFullscreenButton = ce('div', 'entryPainterTopFullscreenButton');
-        painterTopFullscreenButton.setAttribute('title', Lang.Painter.fullscreen);
-        painterTopFullscreenButton.setAttribute('alt', Lang.Painter.fullscreen);
-        painterTopFullscreenButton.addClass('entryPlaygroundPainterFullscreenButton');
-        painterTopFullscreenButton.bindOnClick(() => {
-            this.toggleFullscreen();
-        });
-        this.fullscreenButton = painterTopFullscreenButton;
-        painterTop.appendChild(painterTopFullscreenButton);
-
-        const painterTopMenu = ce('nav', 'entryPainterTopMenu');
-        painterTopMenu.addClass('entryPlaygroundPainterTopMenu');
-        painterTop.appendChild(painterTopMenu);
-
-        const painterTopMenuFileNew = ce('div', 'entryPainterTopMenuFileNew');
-        painterTopMenuFileNew.bindOnClick(painter.newPicture.bind(this));
-        painterTopMenuFileNew.addClass('entryPlaygroundPainterTopMenuFileNew');
-        painterTopMenuFileNew.innerHTML = Lang.Workspace.new_picture;
-        painterTopMenu.appendChild(painterTopMenuFileNew);
-
-        const painterTopMenuFile = ce('div', 'entryPainterTopMenuFile');
-        painterTopMenuFile.addClass('entryPlaygroundPainterTopMenuFile painterTopHeader');
-        painterTopMenuFile.innerHTML = Lang.Menus.offline_file;
-        const painterTopMenuFileDropdown = ce('div');
-
-        painterTopMenuFileDropdown.addClass('entryPlaygroundPainterTopMenuFileDropdown');
-        painterTopMenu.appendChild(painterTopMenuFile);
-        painterTopMenuFile.appendChild(painterTopMenuFileDropdown);
-
-        const painterTopMenuEdit = ce('div', 'entryPainterTopMenuEdit');
-        painterTopMenuEdit.addClass('entryPlaygroundPainterTopMenuEdit painterTopHeader');
-        painterTopMenuEdit.innerHTML = Lang.Menus.offline_edit;
-        painterTopMenu.appendChild(painterTopMenuEdit);
-
-        const painterTopMenuFileSave = ce('div', 'entryPainterTopMenuFileSave');
-        painterTopMenuFileSave.bindOnClick(function() {
-            painter.file_save(false);
-        });
-        painterTopMenuFileSave.addClass('entryPainterTopMenuFileSave');
-        painterTopMenuFileSave.innerHTML = Lang.Workspace.painter_file_save;
-        painterTopMenuFileDropdown.appendChild(painterTopMenuFileSave);
-
-        const painterTopMenuFileSaveAsLink = ce('div', 'entryPainterTopMenuFileSaveAs');
-        painterTopMenuFileSaveAsLink.bindOnClick(function() {
-            painter.file.mode = 'new';
-            painter.file_save(false);
-        });
-        painterTopMenuFileSaveAsLink.addClass('entryPlaygroundPainterTopMenuFileSaveAs');
-        painterTopMenuFileSaveAsLink.innerHTML = Lang.Workspace.painter_file_saveas;
-        painterTopMenuFileDropdown.appendChild(painterTopMenuFileSaveAsLink);
-
-        const painterTopMenuEditDropdown = ce('div');
-        painterTopMenuEditDropdown.addClass('entryPlaygroundPainterTopMenuEditDropdown');
-        painterTopMenuEdit.appendChild(painterTopMenuEditDropdown);
-
-        const painterTopMenuEditImport = ce('div', 'entryPainterTopMenuEditImport');
-        painterTopMenuEditImport.bindOnClick(function() {
-            Entry.dispatchEvent('openPictureImport');
-        });
-        painterTopMenuEditImport.addClass('entryPainterTopMenuEditImport');
-        painterTopMenuEditImport.innerHTML = Lang.Workspace.get_file;
-        painterTopMenuEditDropdown.appendChild(painterTopMenuEditImport);
-
-        const painterTopMenuEditCopy = ce('div', 'entryPainterTopMenuEditCopy');
-        painterTopMenuEditCopy.bindOnClick(function() {
-            painter.copy();
-        });
-        painterTopMenuEditCopy.addClass('entryPlaygroundPainterTopMenuEditCopy');
-        painterTopMenuEditCopy.innerHTML = Lang.Workspace.copy_file;
-        painterTopMenuEditDropdown.appendChild(painterTopMenuEditCopy);
-
-        const painterTopMenuEditCut = ce('div', 'entryPainterTopMenuEditCut');
-        painterTopMenuEditCut.bindOnClick(function() {
-            painter.cut();
-        });
-        painterTopMenuEditCut.addClass('entryPlaygroundPainterTopMenuEditCut');
-        painterTopMenuEditCut.innerHTML = Lang.Workspace.cut_picture;
-        painterTopMenuEditDropdown.appendChild(painterTopMenuEditCut);
-
-        const painterTopMenuEditPaste = ce('div', 'entryPainterTopMenuEditPaste');
-        painterTopMenuEditPaste.bindOnClick(function() {
-            painter.paste();
-        });
-        painterTopMenuEditPaste.addClass('entryPlaygroundPainterTopMenuEditPaste');
-        painterTopMenuEditPaste.innerHTML = Lang.Workspace.paste_picture;
-        painterTopMenuEditDropdown.appendChild(painterTopMenuEditPaste);
-
-        const painterTopMenuEditEraseAll = ce('div', 'entryPainterTopMenuEditEraseAll');
-        painterTopMenuEditEraseAll.addClass('entryPlaygroundPainterTopMenuEditEraseAll');
-        painterTopMenuEditEraseAll.innerHTML = Lang.Workspace.remove_all;
-        painterTopMenuEditEraseAll.bindOnClick(function() {
-            painter.lc.clear();
-        });
-        painterTopMenuEditDropdown.appendChild(painterTopMenuEditEraseAll);
-
-        $(painterTopMenuFile).on('click tab', menuClickEvent);
-        $(painterTopMenuEdit).on('click tab', menuClickEvent);
-        $(document).on('click tap', (e) => {
-            e.stopPropagation();
-            $(painterTopMenuFile).removeClass('active');
-            $(painterTopMenuEdit).removeClass('active');
-        });
-
-        const painterTopStageXY = ce('div', 'entryPainterTopStageXY');
-        const entryPainterTopStageXYLabel = ce('span', 'entryPainterTopStageXYLabel');
-        this.painterTopStageXY = entryPainterTopStageXYLabel;
-        painterTopStageXY.addClass('entryPlaygroundPainterTopStageXY');
-        entryPainterTopStageXYLabel.addClass('entryPainterTopStageXYLabel');
-        painterTopStageXY.appendChild(entryPainterTopStageXYLabel);
-        painterTop.appendChild(painterTopStageXY);
-
-        Entry.addEventListener('pictureSelected', this.changePicture.bind(this));
-
-        function menuClickEvent(e) {
-            $(painterTopMenuFile).removeClass('active');
-            $(painterTopMenuEdit).removeClass('active');
-            if (e.target === this) {
-                e.stopImmediatePropagation();
-                $(this).addClass('active');
-            }
-        }
-    };
-
-    p.stagemousemove = function(event) {
-        this.painterTopStageXY.textContent = `x:${event.x.toFixed(1)}, y:${event.y.toFixed(1)}`;
-    };
-
-    p.attachKeyboardEvents = function() {
+    attachKeyboardEvents() {
         this.detachKeyboardEvents();
 
         const events = this._keyboardEvents;
@@ -476,9 +245,9 @@ Entry.Painter = function(view) {
 
         evt = Entry.keyUpped;
         evt && events.push(evt.attach(this, this._keyboardUpControl));
-    };
+    }
 
-    p.detachKeyboardEvents = function() {
+    detachKeyboardEvents() {
         const events = this._keyboardEvents;
         if (!events || !events.length) {
             return;
@@ -488,9 +257,40 @@ Entry.Painter = function(view) {
             const evt = events.pop();
             evt.destroy && evt.destroy();
         }
-    };
+    }
 
-    p.clear = function() {
+    _keyboardPressControl(e) {}
+
+    toggleFullscreen(isFullscreen) {
+        const { pictureView_ } = Entry.playground;
+        const $view = $(this.view);
+        if ((isFullscreen !== true && $view.hasClass('fullscreen')) || isFullscreen === false) {
+            pictureView_.appendChild(this.view);
+            $view.removeClass('fullscreen');
+            if (this.fullscreenButton) {
+                this.fullscreenButton.setAttribute('title', Lang.Painter.fullscreen);
+                this.fullscreenButton.setAttribute('alt', Lang.Painter.fullscreen);
+            }
+        } else {
+            document.body.appendChild(this.view);
+            $view.addClass('fullscreen');
+            if (this.fullscreenButton) {
+                this.fullscreenButton.setAttribute('title', Lang.Painter.exit_fullscreen);
+                this.fullscreenButton.setAttribute('alt', Lang.Painter.exit_fullscreen);
+            }
+        }
+        this.entryPaint && this.entryPaint.realign();
+    }
+
+    clear() {
         this.toggleFullscreen(false);
-    };
-})(Entry.Painter.prototype);
+    }
+
+    undo() {
+        this.entryPaint.undo();
+    }
+
+    redo() {
+        this.entryPaint.redo();
+    }
+};
