@@ -12,70 +12,68 @@ import { clog } from '../utils/logs';
 import { PrimitiveSet } from './structure/PrimitiveSet';
 import { TimeoutTimer } from '../utils/TimeoutTimer';
 
-
 export class SceneTextures implements ISceneTextures {
+    private _path_tex_map: PrimitiveMap<EntryTexture> = new PrimitiveMap();
+    private _activated: boolean;
+    private _gcTimer: TimeoutTimer = new TimeoutTimer();
 
+    constructor(
+        public sceneID: string,
+        private _option: EntryTextureOption,
+        private _loader: AtlasImageLoader
+    ) {}
 
-    private _path_tex_map:PrimitiveMap<EntryTexture> = new PrimitiveMap();
-    private _activated:boolean;
-    private _gcTimer:TimeoutTimer = new TimeoutTimer();
-
-
-    constructor(public sceneID:string, private _option:EntryTextureOption, private _loader:AtlasImageLoader) {
-
+    _internal_imageRemoved(): void {
+        if (this._gcTimer.isRunning) return;
+        this._gcTimer.timeout(500, () => {
+            this._gcTexture();
+        });
     }
 
-    _internal_imageRemoved():void {
-        if(this._gcTimer.isRunning) return;
-        this._gcTimer.timeout(500, ()=>{ this._gcTexture(); });
-    }
-
-    _gcTexture():void {
-        let usedPathSet:PrimitiveSet = PIXIAtlasHelper.getScenePathSet(this.sceneID);
+    _gcTexture(): void {
+        let usedPathSet: PrimitiveSet = PIXIAtlasHelper.getScenePathSet(this.sceneID);
         let deleteCnt = 0;
-        this._path_tex_map.each((tex:EntryTexture, path:string)=>{
-            if(usedPathSet.hasValue(path)) return;
+        this._path_tex_map.each((tex: EntryTexture, path: string) => {
+            if (usedPathSet.hasValue(path)) return;
             tex.destroy(true);
             this._path_tex_map.remove(path);
             deleteCnt++;
         });
-        if(deleteCnt) {
+        if (deleteCnt) {
             clog(`[SceneTextures] ${deleteCnt} textures deleted`);
         }
     }
 
-    activate():void {
+    activate(): void {
         this._activated = true;
-        this._path_tex_map.each((tex:EntryTexture, path:string)=>{
+        this._path_tex_map.each((tex: EntryTexture, path: string) => {
             let info = this._loader.getImageInfo(path);
-            if(!info || !info.isReady ) {
+            if (!info || !info.isReady) {
                 return;
             }
             this.putImage(info, false);
         });
-
     }
 
-    addPicInfo(pic:IRawPicture):void {
+    addPicInfo(pic: IRawPicture): void {
         let path = PIXIAtlasHelper.getRawPath(pic);
         let map = this._path_tex_map;
-        if(map.hasValue(path)) return;
+        if (map.hasValue(path)) return;
 
         let info = this._loader.getImageInfo(path);
-        let rect:ImageRect = PIXIAtlasHelper.getNewImageRect(pic, this._option.texMaxRect);
-        if(!info) {
+        let rect: ImageRect = PIXIAtlasHelper.getNewImageRect(pic, this._option.texMaxRect);
+        if (!info) {
             this._loader.load(pic, rect);
         }
         let tex = this._newTexture(path, rect);
         map.add(path, tex);
-        if(info&&info.isReady) {
+        if (info && info.isReady) {
             this.putImage(info, false);
         }
     }
 
-
-    private _newTexture(path:string, rect:ImageRect):EntryTexture {
-        let baseTex:EntryBaseTexture = new EntryBaseTexture();
+    private _newTexture(path: string, rect: ImageRect): EntryTexture {
+        let baseTex: EntryBaseTexture = new EntryBaseTexture();
         baseTex.width = rect.width;
         baseTex.height = rect.height;
         baseTex.mipmap = this._option.mipmap;
@@ -85,42 +83,46 @@ export class SceneTextures implements ISceneTextures {
         return tex;
     }
 
-
-    deactivate():void {
+    deactivate(): void {
         this._activated = false;
-        this._path_tex_map.each((tex:EntryTexture, path:string)=>{
+        this._path_tex_map.each((tex: EntryTexture, path: string) => {
             tex.getBaseTexture().dispose();
         });
     }
 
-
-    getTexture(path:string):EntryTexture {
+    getTexture(path: string): EntryTexture {
         return this._path_tex_map.getValue(path);
     }
 
-
-    putImage(info:AtlasImageLoadingInfo, forceUpdateBaseTexture:boolean):void {
-        let tex:EntryTexture = this._path_tex_map.getValue(info.path);
-        if(!tex) {
+    putImage(info: AtlasImageLoadingInfo, forceUpdateBaseTexture: boolean): void {
+        let tex: EntryTexture = this._path_tex_map.getValue(info.path);
+        if (!tex) {
             return;
         }
         let baseTex = tex.getBaseTexture();
-        baseTex.updateSource(info.source());
+        const source = info.source();
+        if (tex._frame) {
+            if (tex._frame.height != info.srcHeight) {
+                tex._frame.height = info.srcHeight;
+            }
+            if (tex._frame.width != info.srcWidth) {
+                tex._frame.width = info.srcWidth;
+            }
+        }
+        baseTex.updateSource(source);
+
         Entry.requestUpdate = true;
     }
 
-
-    destroy():void {
-        clog("destroy",this.sceneID);
+    destroy(): void {
+        clog('destroy', this.sceneID);
         this._gcTimer.reset();
         this._gcTimer = null;
 
-        this._path_tex_map.each((tex:EntryTexture, path:string)=>{
+        this._path_tex_map.each((tex: EntryTexture, path: string) => {
             tex.destroy(true);
         });
         this._path_tex_map.destroy();
         this._path_tex_map = null;
     }
-
-
 }
