@@ -3,12 +3,11 @@ import { EntryBaseTexture } from './EntryBaseTexture';
 import { EntryTextureBase } from './EntryTextureBase';
 
 export class EntryTexture extends EntryTextureBase {
-
-    constructor(baseTexture: EntryBaseTexture, imageRect:ImageRect) {
+    constructor(baseTexture: EntryBaseTexture, imageRect: ImageRect) {
         super(baseTexture, imageRect);
     }
 
-    getBaseTexture():EntryBaseTexture {
+    getBaseTexture(): EntryBaseTexture {
         return this.baseTexture as EntryBaseTexture;
     }
 }
