@@ -8,6 +8,7 @@ import { Backpack, ColorPicker, Dropdown, Sortable } from '@entrylabs/tool';
 import Toast from '../playground/toast';
 import EntryEvent from '@entrylabs/event';
 import { Destroyer } from '../util/destroyer/Destroyer';
+import { saveAs } from 'file-saver';
 
 const Entry = require('../entry');
 
@@ -16,7 +17,7 @@ const Entry = require('../entry');
  * This manage all view related with block.
  * @constructor
  */
-Entry.Playground = class {
+Entry.Playground = class Playground {
     constructor() {
         this._destroyer = this._destroyer || new Destroyer();
         this._destroyer.destroy();
@@ -224,8 +225,8 @@ Entry.Playground = class {
             const backPackButton = Entry.createElement('div')
                 .addClass('entryPlaygroundBackPackButtonWorkspace')
                 .appendTo(tabButtonView);
-            backPackButton.setAttribute('alt', Lang.Blocks.show_all_comment);
-            backPackButton.setAttribute('title', Lang.Blocks.show_all_comment);
+            backPackButton.setAttribute('alt', Lang.Workspace.my_storage);
+            backPackButton.setAttribute('title', Lang.Workspace.my_storage);
 
             this.backPackButton_ = backPackButton;
             backPackButton.bindOnClick(() => {
@@ -597,11 +598,18 @@ Entry.Playground = class {
                 .addClass('entryPlaygroundPictureList')
                 .appendTo(PictureView);
 
-            this.painter = new Entry.Painter(
-                Entry.createElement('div', 'entryPainter')
-                    .addClass('entryPlaygroundPainter')
-                    .appendTo(PictureView)
-            );
+            const painterDom = Entry.createElement('div', 'entryPainter')
+                .addClass('entryPlaygroundPainter')
+                .appendTo(PictureView);
+
+            switch (Entry.paintMode) {
+                case 'entry-paint':
+                    this.painter = new Entry.Painter(painterDom);
+                    break;
+                case 'literallycanvas':
+                    this.painter = new Entry.LiterallycanvasPainter(painterDom);
+                    break;
+            }
         }
     }
 
@@ -666,7 +674,7 @@ Entry.Playground = class {
         fontLink.bindOnClick(() => {
             const options = EntryStatic.fonts
                 .filter((font) => font.visible)
-                .map((font) => [font.name, font]);
+                .map((font) => [font.name, font, font.style]);
             fontLink.addClass('imico_pop_select_arr_up');
             fontLink.removeClass('imico_pop_select_arr_down');
             this.openDropDown(
@@ -678,12 +686,18 @@ Entry.Playground = class {
                     if (that.object.entity.getLineBreak()) {
                         textValue = textEditArea.value;
                     }
-
-                    if (/[\u4E00-\u9FFF]/.exec(textValue) != null) {
-                        font = options[0][1];
-                        entrylms.alert(Lang.Menus.not_supported_text);
+                    const { options = {} } = Entry;
+                    const { textOptions = {} } = options;
+                    const { hanjaEnable } = textOptions;
+                    if (!hanjaEnable) {
+                        if (/[\u4E00-\u9FFF]/.exec(textValue) != null) {
+                            font = options[0][1];
+                            entrylms.alert(Lang.Menus.not_supported_text);
+                        }
                     }
                     fontLink.innerText = font.name;
+                    this.textEditArea.style.fontFamily = font.family;
+                    this.textEditInput.style.fontFamily = font.family;
                     $('#entryTextBoxAttrFontName').data('font', font);
                     this.object.entity.setFontType(font.family);
                 },
@@ -882,7 +896,13 @@ Entry.Playground = class {
             const entity = object.entity;
             const selected = $('#entryTextBoxAttrFontName').data('font');
             const defaultFont = EntryStatic.fonts[0];
-            if (selected.family === 'Nanum Pen Script' || selected.family === 'Jeju Hallasan') {
+            const { options = {} } = Entry;
+            const { textOptions = {} } = options;
+            const { hanjaEnable } = textOptions;
+            if (
+                !hanjaEnable &&
+                (selected.family === 'Nanum Pen Script' || selected.family === 'Jeju Hallasan')
+            ) {
                 if (/[\u4E00-\u9FFF]/.exec(this.value) != null) {
                     $('#entryTextBoxAttrFontName').text(defaultFont.name);
                     entity.setFontType(defaultFont.family);
@@ -1134,7 +1154,7 @@ Entry.Playground = class {
     /**
      * Inject picture
      */
-    injectPicture() {
+    injectPicture(isSelect = true) {
         const view = this.pictureListView_;
         if (!view) {
             return;
@@ -1151,7 +1171,7 @@ Entry.Playground = class {
                 element.orderHolder.innerHTML = i + 1;
             });
 
-            this.selectPicture(this.object.selectedPicture);
+            isSelect && this.selectPicture(this.object.selectedPicture);
         }
 
         this.updatePictureView();
@@ -1161,7 +1181,7 @@ Entry.Playground = class {
      * Add picture
      * @param {picture model} picture
      */
-    addPicture(picture, isNew) {
+    addPicture(picture, isNew, isSelect = true) {
         const tempPicture = _.clone(picture);
 
         if (isNew === true) {
@@ -1178,9 +1198,7 @@ Entry.Playground = class {
 
         this.generatePictureElement(picture);
 
-        Entry.do('objectAddPicture', picture.objectId || this.object.id, picture);
-        this.injectPicture();
-        this.selectPicture(picture);
+        Entry.do('objectAddPicture', picture.objectId || this.object.id, picture, isSelect);
     }
 
     /**
@@ -1221,18 +1239,25 @@ Entry.Playground = class {
      */
     downloadPicture(pictureId) {
         const picture = Entry.playground.object.getPicture(pictureId);
-        if (picture.fileurl) {
-            window.open(
-                `/api/sprite/download/entryjs/${btoa(picture.fileurl)}/${encodeURIComponent(
-                    picture.name
-                )}.png`
-            );
-        } else {
-            window.open(
-                `/api/sprite/download/image/${btoa(picture.filename)}/${encodeURIComponent(
-                    picture.name
-                )}.png`
-            );
+        const { imageType = 'png' } = picture;
+        /**
+            Logic in try phrase will be disregarded after renewal.
+            nt11576
+        */
+        try {
+            if (picture.fileurl) {
+                saveAs(
+                    `/api/sprite/download/entryjs/${btoa(picture.fileurl)}/${encodeURIComponent(
+                        picture.name
+                    )}.png`,
+                    `${picture.name}.${imageType}`
+                );
+            } else {
+                const src = this.painter.getImageSrc(picture);
+                saveAs(src, `${picture.name}.${imageType}`);
+            }
+        } catch (e) {
+            Entry.dispatchEvent('downloadPicture', picture);
         }
     }
 
@@ -1249,7 +1274,7 @@ Entry.Playground = class {
      * Select picture
      * @param {picture}
      */
-    selectPicture(picture) {
+    selectPicture(picture, removed) {
         const pictures = this.object.pictures;
         for (let i = 0, len = pictures.length; i < len; i++) {
             const target = pictures[i];
@@ -1270,7 +1295,7 @@ Entry.Playground = class {
             if (!picture.objectId) {
                 picture.objectId = this.object.id;
             }
-            Entry.dispatchEvent('pictureSelected', picture);
+            Entry.dispatchEvent('pictureSelected', picture, removed);
         }
     }
 
@@ -1418,6 +1443,8 @@ Entry.Playground = class {
             } else {
                 window.open(sound.fileurl);
             }
+        } else if (sound.path.indexOf('sound') > -1) {
+            Entry.dispatchEvent('downloadSound', sound);
         } else {
             window.open(
                 `/api/sprite/download/sound/${encodeURIComponent(
@@ -1662,6 +1689,54 @@ Entry.Playground = class {
         }
     }
 
+    nameViewBlur() {
+        if (!Entry.playground.nameViewFocus) {
+            return;
+        }
+        if (this.nameView.value.trim() === '') {
+            entrylms.alert(Lang.Workspace.enter_the_name).on('hide', () => {
+                this.nameView.focus();
+            });
+            return true;
+        }
+
+        let nameViewArray = $('.entryPlaygroundPictureName');
+        if (nameViewArray.length !== Entry.playground.object.pictures.length) {
+            nameViewArray = nameViewArray.slice(0, -1); // pop last element (드래그 시 발생하는 임시 엘리먼트임)
+        }
+
+        for (let i = 0; i < nameViewArray.length; i++) {
+            if (
+                nameViewArray.eq(i).val() == this.nameView.value &&
+                nameViewArray[i] != this.nameView
+            ) {
+                entrylms.alert(Lang.Workspace.name_already_exists).on('hide', () => {
+                    this.nameView.focus();
+                });
+                return true;
+            }
+        }
+        const newValue = this.nameView.value;
+        this.nameView.picture.name = newValue;
+        const playground = Entry.playground;
+        if (playground) {
+            if (playground.object) {
+                const pic = playground.object.getPicture(this.nameView.picture.id);
+                if (pic) {
+                    pic.name = newValue;
+                }
+            }
+            const painter = playground.painter;
+            if (painter && painter.file) {
+                painter.file.name = newValue;
+            }
+
+            playground.reloadPlayground();
+        }
+        Entry.dispatchEvent('pictureNameChanged', this.nameView.picture);
+        Entry.playground.nameViewFocus = false;
+    }
+
     generatePictureElement(picture) {
         const element = Entry.createElement('li', picture.id)
             .addClass('entryPlaygroundPictureElement')
@@ -1689,19 +1764,7 @@ Entry.Playground = class {
                 {
                     text: Lang.Workspace.context_remove,
                     callback() {
-                        if (Entry.playground.object.removePicture(picture.id)) {
-                            Entry.removeElement(element);
-                            Entry.dispatchEvent('removePicture', picture);
-                            Entry.toast.success(
-                                Lang.Workspace.shape_remove_ok,
-                                `${picture.name} ${Lang.Workspace.shape_remove_ok_msg}`
-                            );
-                        } else {
-                            Entry.toast.alert(
-                                Lang.Workspace.shape_remove_fail,
-                                Lang.Workspace.shape_remove_fail_msg
-                            );
-                        }
+                        Entry.playground._removePicture(picture, element);
                     },
                 },
                 {
@@ -1744,74 +1807,28 @@ Entry.Playground = class {
             .addClass('entryEllipsis');
         nameView.picture = picture;
         nameView.value = picture.name;
-        Entry.attachEventListener(nameView, 'blur', nameViewBlur);
-
-        function nameViewBlur() {
-            if (this.value.trim() === '') {
-                return entrylms.alert(Lang.Workspace.enter_the_name).on('hide', () => {
-                    nameView.focus();
-                });
-            }
-
-            let nameViewArray = $('.entryPlaygroundPictureName');
-            if (nameViewArray.length !== Entry.playground.object.pictures.length) {
-                nameViewArray = nameViewArray.slice(0, -1); // pop last element (드래그 시 발생하는 임시 엘리먼트임)
-            }
-
-            for (let i = 0; i < nameViewArray.length; i++) {
-                if (nameViewArray.eq(i).val() == nameView.value && nameViewArray[i] != this) {
-                    return entrylms.alert(Lang.Workspace.name_already_exists).on('hide', () => {
-                        nameView.focus();
-                    });
-                }
-            }
-            const newValue = this.value;
-            this.picture.name = newValue;
-            const playground = Entry.playground;
-            if (playground) {
-                if (playground.object) {
-                    const pic = playground.object.getPicture(this.picture.id);
-                    if (pic) {
-                        pic.name = newValue;
-                    }
-                }
-                const painter = playground.painter;
-                if (painter && painter.file) {
-                    painter.file.name = newValue;
-                }
-
-                playground.reloadPlayground();
-            }
-            Entry.dispatchEvent('pictureNameChanged', this.picture);
-        }
+        Entry.attachEventListener(nameView, 'blur', this.nameViewBlur.bind(this));
+        Entry.attachEventListener(nameView, 'focus', (e) => {
+            this.nameView = e.target;
+            this.nameViewFocus = true;
+        });
 
         nameView.onkeypress = Entry.Utils.blurWhenEnter;
         element.appendChild(nameView);
         Entry.createElement('div', `s_${picture.id}`)
             .addClass('entryPlaygroundPictureSize')
-            .appendTo(element).innerHTML = `${picture.dimension.width} X ${
-            picture.dimension.height
-        }`;
+            .appendTo(
+                element
+            ).innerHTML = `${picture.dimension.width} X ${picture.dimension.height}`;
 
         const removeButton = Entry.createElement('div').addClass('entryPlayground_del');
         const { Buttons = {} } = Lang || {};
         const { delete: delText = '삭제' } = Buttons;
         removeButton.appendTo(element).innerText = delText;
-        removeButton.bindOnClick(() => {
+        removeButton.bindOnClick((e) => {
             try {
-                if (Entry.playground.object.removePicture(picture.id)) {
-                    Entry.removeElement(element);
-                    Entry.dispatchEvent('removePicture', picture);
-                    Entry.toast.success(
-                        Lang.Workspace.shape_remove_ok,
-                        `${picture.name} ${Lang.Workspace.shape_remove_ok_msg}`
-                    );
-                } else {
-                    Entry.toast.alert(
-                        Lang.Workspace.shape_remove_fail,
-                        Lang.Workspace.shape_remove_fail_msg
-                    );
-                }
+                e.stopPropagation();
+                this._removePicture(picture, element);
             } catch (e) {
                 Entry.toast.alert(
                     Lang.Workspace.shape_remove_fail,
@@ -1819,6 +1836,22 @@ Entry.Playground = class {
                 );
             }
         });
+    }
+
+    _removePicture(picture, element) {
+        if (Entry.playground.object.pictures.length > 1) {
+            Entry.do('objectRemovePicture', picture.objectId, picture);
+            Entry.removeElement(element);
+            Entry.toast.success(
+                Lang.Workspace.shape_remove_ok,
+                `${picture.name} ${Lang.Workspace.shape_remove_ok_msg}`
+            );
+        } else {
+            Entry.toast.alert(
+                Lang.Workspace.shape_remove_fail,
+                Lang.Workspace.shape_remove_fail_msg
+            );
+        }
     }
 
     generateSoundElement(sound) {
@@ -2012,17 +2045,22 @@ Entry.Playground = class {
     };
 
     openColourPicker = (target, color, canTransparent, callback) => {
+        const containers = $('.entry-color-picker');
+        if (containers.length > 0) {
+            $(target).removeClass('on');
+            return containers.remove();
+        }
         const container = Entry.Dom('div', {
             class: 'entry-color-picker',
             parent: $('body'),
         })[0];
-
         $(target).addClass('on');
         const colorPicker = new ColorPicker({
             data: {
                 color,
                 positionDom: target,
                 canTransparent,
+                outsideExcludeDom: [target],
                 onOutsideClick: (color) => {
                     if (colorPicker) {
                         $(target).removeClass('on');
