@@ -12,7 +12,9 @@ Entry.Popup = function(className) {
 
     this.body_ = Entry.createElement('div');
     this.body_.addClass('entryPopup');
-    if (className) this.body_.addClass(className);
+    if (className) {
+        this.body_.addClass(className);
+    }
     this.body_.bindOnClick(function(e) {
         if (e.target == this) {
             this.popup.remove();
@@ -22,9 +24,11 @@ Entry.Popup = function(className) {
     document.body.appendChild(this.body_);
     this.window_ = Entry.createElement('div');
     this.window_.addClass('entryPopupWindow');
-    if (Entry.targetChecker && !Entry.targetChecker.statusViewDisabled) this.window_.addClass('targetCheckerPopup');
+    if (Entry.targetChecker && !Entry.targetChecker.statusViewDisabled) {
+        this.window_.addClass('targetCheckerPopup');
+    }
     // if (Entry.device === 'tablet') this.window_.addClass('tablet');
-    this.window_.bindOnClick(function() {});
+    this.window_.bindOnClick(() => {});
     Entry.addEventListener('windowResized', this.resize);
     window.popup = this;
     this.resize();
@@ -37,7 +41,6 @@ Entry.Popup = function(className) {
 Entry.Popup.prototype.remove = function() {
     while (this.window_.hasChildNodes()) {
         if (Entry.type == 'workspace') {
-            console.log(this.window_.firstChild, Entry.container.view_);
             Entry.engineContainer.insertBefore(
                 this.window_.firstChild,
                 Entry.engineContainer.firstChild
@@ -57,8 +60,13 @@ Entry.Popup.prototype.remove = function() {
     Entry.removeEventListener('windowResized', this.resize);
     Entry.engine.popup = null;
     Entry.windowResized.notify();
-    if (Entry.type === 'workspace' && Entry.targetChecker && !Entry.targetChecker.statusViewDisabled)
+    if (
+        Entry.type === 'workspace' &&
+        Entry.targetChecker &&
+        !Entry.targetChecker.statusViewDisabled
+    ) {
         Entry.targetChecker.getStatusView().remove();
+    }
 };
 
 /**
@@ -66,21 +74,22 @@ Entry.Popup.prototype.remove = function() {
  * @param {event} e
  */
 Entry.Popup.prototype.resize = function(e) {
-    var popup = window.popup;
-    var popupWindow = popup.window_;
-    var bottomOffset = Entry.targetChecker && !Entry.targetChecker.statusViewDisabled ? 91 + 48 : 48;
-    var maxWidth = window.innerWidth * 0.9;
-    var maxHeight = window.innerHeight * 0.9 - bottomOffset;
+    const popup = window.popup;
+    const popupWindow = popup.window_;
+    const bottomOffset =
+        Entry.targetChecker && !Entry.targetChecker.statusViewDisabled ? 91 + 48 : 48;
+    let maxWidth = window.innerWidth * 0.9;
+    let maxHeight = window.innerHeight * 0.9 - bottomOffset;
     if (maxWidth * 9 <= maxHeight * 16) {
-        maxHeight = maxWidth / 16 * 9;
+        maxHeight = (maxWidth / 16) * 9;
         maxHeight += bottomOffset;
-        popupWindow.style.width = String(maxWidth) + 'px';
-        popupWindow.style.height = String(maxHeight) + 'px';
+        popupWindow.style.width = `${String(maxWidth)}px`;
+        popupWindow.style.height = `${String(maxHeight)}px`;
     } else {
-        maxWidth = maxHeight * 16 / 9;
+        maxWidth = (maxHeight * 16) / 9;
         maxHeight += bottomOffset;
-        popupWindow.style.width = String(maxWidth) + 'px';
-        popupWindow.style.height = String(maxHeight) + 'px';
+        popupWindow.style.width = `${String(maxWidth)}px`;
+        popupWindow.style.height = `${String(maxHeight)}px`;
     }
 
     Entry.stage && Entry.stage.updateBoundRect();
