@@ -3,8 +3,6 @@
  */
 'use strict';
 
-
-
 /**
  * Constructor of toast
  * @constructor
@@ -12,106 +10,117 @@
 Entry.Toast = function() {
     this.toasts_ = [];
     /** @type {Element} */
-    var exist = document.getElementById('entryToastContainer');
-    if (exist)
+    const exist = document.getElementById('entryToastContainer');
+    if (exist) {
         document.body.removeChild(exist);
+    }
     this.body_ = Entry.createElement('div', 'entryToastContainer');
     this.body_.addClass('entryToastContainer');
     document.body.appendChild(this.body_);
 };
 
 Entry.Toast.prototype.warning = function(title, message, isNotAutoDispose) {
-    var toast = Entry.createElement('div', 'entryToast');
+    const toast = Entry.createElement('div', 'entryToast');
     toast.addClass('entryToast');
     toast.addClass('entryToastWarning');
-    toast.bindOnClick(function() {Entry.toast.body_.removeChild(this);});
-    var toastTitle = Entry.createElement('div', 'entryToast');
+    toast.bindOnClick(function() {
+        Entry.toast.body_.removeChild(this);
+    });
+    const toastTitle = Entry.createElement('div', 'entryToast');
     toastTitle.addClass('entryToastTitle');
-    toastTitle.innerHTML = title;
+    toastTitle.textContent = title;
     toast.appendChild(toastTitle);
-    var toastMessage = Entry.createElement('p', 'entryToast');
+    const toastMessage = Entry.createElement('p', 'entryToast');
     toastMessage.addClass('entryToastMessage');
-    toastMessage.innerHTML = message;
+    toastMessage.textContent = message;
     toast.appendChild(toastMessage);
     this.toasts_.push(toast);
     this.body_.appendChild(toast);
-    var f = function() {
+    const f = function() {
         toast.style.opacity = 1;
-        var timer = setInterval(function () {
+        var timer = setInterval(() => {
             if (toast.style.opacity < 0.05) {
                 clearInterval(timer);
                 toast.style.display = 'none';
                 Entry.removeElement(toast);
             }
-            toast.style.opacity *= 0.90;
+            toast.style.opacity *= 0.9;
         }, 20);
     };
-    if (!isNotAutoDispose)
+    if (!isNotAutoDispose) {
         window.setTimeout(f, 1000);
+    }
 };
 
 Entry.Toast.prototype.success = function(title, message, isNotAutoDispose) {
-    var toast = Entry.createElement('div', 'entryToast');
+    const toast = Entry.createElement('div', 'entryToast');
     toast.addClass('entryToast');
     toast.addClass('entryToastSuccess');
-    toast.bindOnClick(function() {Entry.toast.body_.removeChild(this);});
-    var toastTitle = Entry.createElement('div', 'entryToast');
+    toast.bindOnClick(function() {
+        Entry.toast.body_.removeChild(this);
+    });
+    const toastTitle = Entry.createElement('div', 'entryToast');
     toastTitle.addClass('entryToastTitle');
-    toastTitle.innerHTML = title;
+    toastTitle.textContent = title;
     toast.appendChild(toastTitle);
-    var toastMessage = Entry.createElement('p', 'entryToast');
+    const toastMessage = Entry.createElement('p', 'entryToast');
     toastMessage.addClass('entryToastMessage');
-    toastMessage.innerHTML = message;
+    toastMessage.textContent = message;
     toast.appendChild(toastMessage);
     this.toasts_.push(toast);
     this.body_.appendChild(toast);
-    var f = function() {
+    const f = function() {
         toast.style.opacity = 1;
-        var timer = setInterval(function () {
+        var timer = setInterval(() => {
             if (toast.style.opacity < 0.05) {
                 clearInterval(timer);
                 toast.style.display = 'none';
                 Entry.removeElement(toast);
             }
-            toast.style.opacity *= 0.90;
+            toast.style.opacity *= 0.9;
         }, 20);
     };
-    if (!isNotAutoDispose)
+    if (!isNotAutoDispose) {
         window.setTimeout(f, 1000);
+    }
 };
 
 Entry.Toast.prototype.alert = function(title, message, isNotAutoDispose) {
-    var toast = Entry.createElement('div', 'entryToast');
-    var timer;
+    const toast = Entry.createElement('div', 'entryToast');
+    let timer;
     toast.addClass('entryToast');
     toast.addClass('entryToastAlert');
     toast.bindOnClick(function() {
         Entry.toast.body_.removeChild(this);
-        if (timer) clearInterval(timer);
+        if (timer) {
+            clearInterval(timer);
+        }
     });
-    var toastTitle = Entry.createElement('div', 'entryToast');
+    const toastTitle = Entry.createElement('div', 'entryToast');
     toastTitle.addClass('entryToastTitle');
-    toastTitle.innerHTML = title;
+    toastTitle.textContent = title;
     toast.appendChild(toastTitle);
-    var toastMessage = Entry.createElement('p', 'entryToast');
+    const toastMessage = Entry.createElement('p', 'entryToast');
     toastMessage.addClass('entryToastMessage');
-    toastMessage.innerHTML = message;
+    toastMessage.textContent = message;
     toast.appendChild(toastMessage);
     this.toasts_.push(toast);
     this.body_.appendChild(toast);
-    var f = function() {
+    const f = function() {
         toast.style.opacity = 1;
-        timer = setInterval(function () {
+        timer = setInterval(() => {
             if (toast.style.opacity < 0.05) {
                 clearInterval(timer);
                 toast.style.display = 'none';
                 //check element already removed from parent
-                if (toast.parentElement)
+                if (toast.parentElement) {
                     Entry.toast.body_.removeChild(toast);
+                }
             }
-            toast.style.opacity *= 0.90;
+            toast.style.opacity *= 0.9;
         }, 20);
     };
-    if (!isNotAutoDispose)
+    if (!isNotAutoDispose) {
         window.setTimeout(f, 5000);
+    }
 };
