@@ -4,7 +4,8 @@
 'use strict';
 
 import SimpleBar from 'simplebar';
-import axios from 'axios';
+import fetch from 'isomorphic-fetch';
+import xssFilters from 'xss-filters';
 
 /**
  * Block variable constructor
@@ -125,19 +126,19 @@ Entry.VariableContainer = class VariableContainer {
     generateAddButtons() {
         const createElement = Entry.createElement;
         const variableAddButton = createElement('button').addClass('entryVariableAddWorkspace');
-        variableAddButton.innerHTML = Lang.Workspace.variable_add;
+        variableAddButton.textContent = Lang.Workspace.variable_add;
         this.variableAddButton_ = variableAddButton;
 
         const messageAddButton = createElement('button').addClass('entryVariableAddWorkspace');
-        messageAddButton.innerHTML = Lang.Workspace.message_create;
+        messageAddButton.textContent = Lang.Workspace.message_create;
         this.messageAddButton_ = messageAddButton;
 
         const listAddButton = createElement('button').addClass('entryVariableAddWorkspace');
-        listAddButton.innerHTML = Lang.Workspace.list_create;
+        listAddButton.textContent = Lang.Workspace.list_create;
         this.listAddButton_ = listAddButton;
 
         const functionAddButton = createElement('button').addClass('entryVariableAddWorkspace');
-        functionAddButton.innerHTML = Lang.Workspace.function_add;
+        functionAddButton.textContent = Lang.Workspace.function_add;
         this.functionAddButton_ = functionAddButton;
     }
 
@@ -276,7 +277,7 @@ Entry.VariableContainer = class VariableContainer {
                 element.appendChild(thumb);
                 Entry.createElement('span')
                     .addClass('text')
-                    .appendTo(element).innerHTML = `${caller.object.name} : ${
+                    .appendTo(element).textContent = `${caller.object.name} : ${
                     Lang.Blocks[`START_${caller.block.type}`]
                 }`;
                 element.bindOnClick((e) => {
@@ -298,7 +299,7 @@ Entry.VariableContainer = class VariableContainer {
             listView.appendChild(fragment);
         } else {
             usedWrapper = Entry.createElement('p').addClass('caution_dsc');
-            usedWrapper.innerHTML = Lang.Workspace.no_use;
+            usedWrapper.textContent = Lang.Workspace.no_use;
         }
         message.usedView = usedWrapper;
         message.listElement.appendChild(usedWrapper);
@@ -320,10 +321,10 @@ Entry.VariableContainer = class VariableContainer {
             .appendTo(usedWrapper);
 
         if (variable.type === 'variable') {
-            usedSubject.innerHTML = Lang.Workspace.Variable_used_objects;
+            usedSubject.textContent = Lang.Workspace.Variable_used_objects;
         } else {
             // variable.type === 'list'
-            usedSubject.innerHTML = Lang.Workspace.List_used_objects;
+            usedSubject.textContent = Lang.Workspace.List_used_objects;
         }
 
         const listView = Entry.createElement('ul')
@@ -341,7 +342,7 @@ Entry.VariableContainer = class VariableContainer {
                 element.appendChild(thumb);
                 Entry.createElement('span')
                     .addClass('text')
-                    .appendTo(element).innerHTML = `${caller.object.name} : ${
+                    .appendTo(element).textContent = `${caller.object.name} : ${
                     Lang.Blocks[`VARIABLE_${caller.block.type}`]
                 }`;
                 element.variable = variable;
@@ -365,7 +366,7 @@ Entry.VariableContainer = class VariableContainer {
         } else {
             Entry.createElement('li')
                 .addClass('text red')
-                .appendTo(listView).innerHTML = Lang.Workspace.no_use;
+                .appendTo(listView).textContent = Lang.Workspace.no_use;
         }
 
         this.variableSettingView && this.variableSettingView.appendChild(usedWrapper);
@@ -396,7 +397,7 @@ Entry.VariableContainer = class VariableContainer {
                 thumb.addClass('thmb');
                 element.appendChild(thumb);
                 const nameElement = Entry.createElement('span').addClass('text');
-                nameElement.innerHTML = caller.object.name;
+                nameElement.textContent = caller.object.name;
                 element.appendChild(nameElement);
                 element.bindOnClick(() => {
                     if (Entry.playground.object != caller.object) {
@@ -414,7 +415,7 @@ Entry.VariableContainer = class VariableContainer {
             listView.appendChild(fragment);
         } else {
             usedWrapper = Entry.createElement('p').addClass('caution_dsc');
-            usedWrapper.innerHTML = Lang.Workspace.no_use;
+            usedWrapper.textContent = Lang.Workspace.no_use;
         }
 
         func.usedView = usedWrapper;
@@ -605,8 +606,8 @@ Entry.VariableContainer = class VariableContainer {
 
         const gLength = (globalV || []).length;
         const lLength = (localV || []).length;
-        globalListTitle.innerHTML = `${Lang.Workspace.Variable_used_at_all_objects} (${gLength})`;
-        localListTitle.innerHTML = `${Lang.Workspace.Variable_used_at_special_object} (${lLength})`;
+        globalListTitle.textContent = `${Lang.Workspace.Variable_used_at_all_objects} (${gLength})`;
+        localListTitle.textContent = `${Lang.Workspace.Variable_used_at_special_object} (${lLength})`;
         this.foldTab(globalList, isGlobalFolded, gLength);
         this.foldTab(localList, isLocalFolded, lLength);
 
@@ -687,8 +688,8 @@ Entry.VariableContainer = class VariableContainer {
 
         const gLength = (globalV || []).length;
         const lLength = (localV || []).length;
-        globalListTitle.innerHTML = `${Lang.Workspace.List_used_all_objects} (${gLength})`;
-        localListTitle.innerHTML = `${Lang.Workspace.list_used_specific_objects} (${lLength})`;
+        globalListTitle.textContent = `${Lang.Workspace.List_used_all_objects} (${gLength})`;
+        localListTitle.textContent = `${Lang.Workspace.list_used_specific_objects} (${lLength})`;
         this.foldTab(globalList, isGlobalFolded, gLength);
         this.foldTab(localList, isLocalFolded, lLength);
 
@@ -1088,7 +1089,7 @@ Entry.VariableContainer = class VariableContainer {
             this.createFunctionView(func);
         }
         if (func.listElement) {
-            func.listElement.nameField.innerHTML = func.description;
+            func.listElement.nameField.textContent = func.description;
         }
 
         this.updateList();
@@ -1118,7 +1119,7 @@ Entry.VariableContainer = class VariableContainer {
         const editBoxInput = createElement('span')
             .addClass('input')
             .appendTo(editBoxInputWrapper);
-        editBoxInput.innerHTML = func.description;
+        editBoxInput.textContent = func.description;
 
         const delButton = createElement('a')
             .addClass('del')
@@ -1799,7 +1800,7 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('Workspace_text')
-            .appendTo(addSpaceGlobalWrapper).innerHTML = Lang.Workspace.use_all_objects;
+            .appendTo(addSpaceGlobalWrapper).textContent = Lang.Workspace.use_all_objects;
 
         createElement('span')
             .addClass('entryVariableAddSpaceCheckWorkspace')
@@ -1824,7 +1825,7 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('entryVariableAddSpaceCloudSpanWorkspace')
-            .appendTo(addSpaceCloudWrapper).innerHTML = Lang.Workspace.Variable_create_cloud;
+            .appendTo(addSpaceCloudWrapper).textContent = Lang.Workspace.Variable_create_cloud;
 
         createElement('span')
             .addClass('entryVariableAddSpaceCheckWorkspace')
@@ -1844,7 +1845,7 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('Workspace_text')
-            .appendTo(addSpaceLocalWrapper).innerHTML = Lang.Workspace.Variable_use_this_object;
+            .appendTo(addSpaceLocalWrapper).textContent = Lang.Workspace.Variable_use_this_object;
 
         createElement('span')
             .addClass('entryVariableAddSpaceCheckWorkspace')
@@ -1864,7 +1865,7 @@ Entry.VariableContainer = class VariableContainer {
             })
             .appendTo(addSpaceButtonWrapper);
         addSpaceCancelButton.href = '#';
-        addSpaceCancelButton.innerHTML = Lang.Buttons.cancel;
+        addSpaceCancelButton.textContent = Lang.Buttons.cancel;
 
         const addSpaceConfirmButton = createElement('a')
             .addClass('entryVariableAddSpaceConfirmWorkspace')
@@ -1874,7 +1875,7 @@ Entry.VariableContainer = class VariableContainer {
             })
             .appendTo(addSpaceButtonWrapper);
         addSpaceConfirmButton.href = '#';
-        addSpaceConfirmButton.innerHTML = Lang.Buttons.save;
+        addSpaceConfirmButton.textContent = Lang.Buttons.save;
         this.variableAddConfirmButton = addSpaceConfirmButton;
     }
 
@@ -1983,7 +1984,7 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('Workspace_text')
-            .appendTo(addSpaceGlobalWrapper).innerHTML = Lang.Workspace.use_all_objects;
+            .appendTo(addSpaceGlobalWrapper).textContent = Lang.Workspace.use_all_objects;
 
         createElement('span')
             .addClass('entryVariableAddSpaceCheckWorkspace')
@@ -2008,7 +2009,7 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('entryVariableAddSpaceCloudSpanWorkspace')
-            .appendTo(addSpaceCloudWrapper).innerHTML = Lang.Workspace.List_create_cloud;
+            .appendTo(addSpaceCloudWrapper).textContent = Lang.Workspace.List_create_cloud;
 
         createElement('span')
             .addClass('entryVariableAddSpaceCheckWorkspace')
@@ -2028,7 +2029,7 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('Workspace_text')
-            .appendTo(addSpaceLocalWrapper).innerHTML = Lang.Workspace.Variable_use_this_object;
+            .appendTo(addSpaceLocalWrapper).textContent = Lang.Workspace.Variable_use_this_object;
 
         createElement('span')
             .addClass('entryVariableAddSpaceCheckWorkspace')
@@ -2048,7 +2049,7 @@ Entry.VariableContainer = class VariableContainer {
             })
             .appendTo(addSpaceButtonWrapper);
         addSpaceCancelButton.href = '#';
-        addSpaceCancelButton.innerHTML = Lang.Buttons.cancel;
+        addSpaceCancelButton.textContent = Lang.Buttons.cancel;
 
         const addSpaceConfirmButton = createElement('a')
             .addClass('entryVariableAddSpaceConfirmWorkspace')
@@ -2058,7 +2059,7 @@ Entry.VariableContainer = class VariableContainer {
             })
             .appendTo(addSpaceButtonWrapper);
         addSpaceConfirmButton.href = '#';
-        addSpaceConfirmButton.innerHTML = Lang.Buttons.save;
+        addSpaceConfirmButton.textContent = Lang.Buttons.save;
         this.listAddConfirmButton = addSpaceConfirmButton;
     }
 
@@ -2108,7 +2109,7 @@ Entry.VariableContainer = class VariableContainer {
             })
             .appendTo(buttonWrapper);
         msgCancel.href = '#';
-        msgCancel.innerHTML = Lang.Buttons.cancel;
+        msgCancel.textContent = Lang.Buttons.cancel;
 
         const msgConfirm = createElement('a')
             .addClass('entryVariableAddSpaceConfirmWorkspace')
@@ -2129,28 +2130,33 @@ Entry.VariableContainer = class VariableContainer {
             })
             .appendTo(buttonWrapper);
         msgConfirm.href = '#';
-        msgConfirm.innerHTML = Lang.Buttons.save;
+        msgConfirm.textContent = Lang.Buttons.save;
 
         msgAddSpace.nameField = msgNameInput;
     }
 
+    /**
+     * 자료형 블록메뉴에서 변수, 리스트 추가 버튼 클릭시 발생하는 함수
+     * 패널을 속성탭으로 옮기고 해당 타입의 컨테이너를 오픈한다.
+     * @param type {'variable'|'list'|'message'}
+     */
     openVariableAddPanel(type = 'variable') {
         Entry.playground.toggleOnVariableView();
         Entry.playground.changeViewMode('variable');
         switch (type) {
             case 'variable':
-                this._getAddPanel().isOpen = true;
                 this.selectFilter(type);
                 this.updateVariableAddView(type);
+                this.clickVariableAddButton(true);
                 break;
             case 'list':
-                this.listAddPanel.isOpen = true;
                 this.selectFilter(type);
                 this.updateVariableAddView(type);
+                this.clickListAddButton(true);
                 break;
             case 'message':
-                this.messageAddPanel.isOpen = true;
                 this.selectFilter(type);
+                this.clickMessageAddButton(true);
                 break;
         }
     }
@@ -2200,6 +2206,7 @@ Entry.VariableContainer = class VariableContainer {
     }
 
     generateTimer(timer) {
+        const x = 240 - (Lang.Workspace.Variable_Timer.length * 12 + 70);
         timer =
             timer ||
             Entry.Variable.create({
@@ -2208,7 +2215,7 @@ Entry.VariableContainer = class VariableContainer {
                 value: 0,
                 variableType: 'timer',
                 visible: false,
-                x: 134,
+                x,
                 y: -70,
             });
 
@@ -2262,7 +2269,7 @@ Entry.VariableContainer = class VariableContainer {
         const boxSubject = createElement('span')
             .addClass('box_sjt')
             .appendTo(varAttr);
-        boxSubject.innerHTML = Lang.Workspace.Variable_property;
+        boxSubject.textContent = Lang.Workspace.Variable_property;
 
         // 기본 값 입력 창
         const attrInputBox = createElement('div')
@@ -2274,7 +2281,7 @@ Entry.VariableContainer = class VariableContainer {
 
         const attrInputLabel = createElement('label').appendTo(attrInputBox);
         attrInputLabel.setAttribute('for', 'attr_cnt');
-        attrInputLabel.innerHTML = Lang.Workspace.default_value;
+        attrInputLabel.textContent = Lang.Workspace.default_value;
 
         const attrInputWrapper = createElement('span')
             .appendTo(attrInputBox)
@@ -2312,7 +2319,7 @@ Entry.VariableContainer = class VariableContainer {
         const slideCheckText = createElement('span')
             .addClass('chk_text')
             .appendTo(slideCheckBox);
-        slideCheckText.innerHTML = Lang.Workspace.slide;
+        slideCheckText.textContent = Lang.Workspace.slide;
 
         // 최소 최대 영역
         const slideCountBox = createElement('div')
@@ -2320,7 +2327,7 @@ Entry.VariableContainer = class VariableContainer {
             .appendTo(slideInputBox);
 
         const minValueInput = createElement('input').appendTo(slideCountBox);
-        minValueInput.innerHTML = Lang.Workspace.min_value;
+        minValueInput.textContent = Lang.Workspace.min_value;
         minValueInput.setAttribute('type', 'text');
 
         const v = that.selected;
@@ -2343,10 +2350,10 @@ Entry.VariableContainer = class VariableContainer {
 
         createElement('span')
             .addClass('dash')
-            .appendTo(slideCountBox).innerHTML = '~';
+            .appendTo(slideCountBox).textContent = '~';
 
         const maxValueInput = createElement('input').appendTo(slideCountBox);
-        maxValueInput.innerHTML = Lang.Workspace.max_value;
+        maxValueInput.textContent = Lang.Workspace.max_value;
         maxValueInput.setAttribute('type', 'text');
 
         if (vType === 'slide') {
@@ -2423,7 +2430,7 @@ Entry.VariableContainer = class VariableContainer {
         const boxSubject = createElement('span')
             .addClass('box_sjt')
             .appendTo(listAttr);
-        boxSubject.innerHTML = Lang.Workspace.list_property;
+        boxSubject.textContent = Lang.Workspace.list_property;
 
         this.generateListImportExportView(listAttr);
         this.generateListCountView(listAttr);
@@ -2452,7 +2459,7 @@ Entry.VariableContainer = class VariableContainer {
                 }
             })
             .appendTo(buttonBox);
-        buttonExport.innerHTML = Lang.Workspace.list_export;
+        buttonExport.textContent = Lang.Workspace.list_export;
 
         const buttonImport = createElement('a')
             .addClass('btn_list')
@@ -2461,7 +2468,7 @@ Entry.VariableContainer = class VariableContainer {
                 Entry.dispatchEvent('openImportListModal');
             })
             .appendTo(buttonBox);
-        buttonImport.innerHTML = Lang.Workspace.list_import;
+        buttonImport.textContent = Lang.Workspace.list_import;
     }
 
     generateListCountView(element) {
@@ -2475,7 +2482,7 @@ Entry.VariableContainer = class VariableContainer {
         const countSubject = createElement('span')
             .addClass('cnt_sjt')
             .appendTo(listCount);
-        countSubject.innerHTML = Lang.Workspace.number_of_list;
+        countSubject.textContent = Lang.Workspace.number_of_list;
 
         const countInputBox = createElement('div')
             .addClass('cnt_inpt')
@@ -2490,7 +2497,7 @@ Entry.VariableContainer = class VariableContainer {
                 Entry.do('listChangeLength', id_, 'minus');
             })
             .appendTo(countInputBox);
-        buttonMinus.innerHTML = '-';
+        buttonMinus.textContent = '-';
         buttonMinus.href = '#';
         this.listSettingView.minus = buttonMinus;
 
@@ -2521,7 +2528,7 @@ Entry.VariableContainer = class VariableContainer {
                 }
             })
             .appendTo(countInputBox);
-        buttonPlus.innerHTML = '+';
+        buttonPlus.textContent = '+';
         buttonPlus.href = '#';
         this.listSettingView.plus = buttonPlus;
 
@@ -2593,14 +2600,14 @@ Entry.VariableContainer = class VariableContainer {
             const fragment = document.createDocumentFragment();
             Entry.createElement('p')
                 .addClass('caution_dsc')
-                .appendTo(fragment).innerHTML = Lang.Workspace.empty_of_list;
+                .appendTo(fragment).textContent = Lang.Workspace.empty_of_list;
             listValues.appendChild(fragment);
         } else {
             const data = arr.map(({ data: value }, i) =>
                 /* html */ `
                     <li>
                         <span class='cnt'>${i + startIndex}</span>
-                        <input value='${value}' type='text' data-index='${i}'/>
+                        <input value='${xssFilters.inSingleQuotedAttr(value)}' type='text' data-index='${i}'/>
                         <a class='del' data-index='${i}'></a>
                     </li>`.trim()
             );
@@ -2716,9 +2723,22 @@ Entry.VariableContainer = class VariableContainer {
             return;
         }
 
-        Entry.dispatchEvent('updateCloudVariable', {
-            variables,
-            lists,
+        let csrfToken = '';
+        try {
+            csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
+        } catch (e) {}
+
+        fetch(`/api/project/variable/${projectId}`, {
+            method: 'PUT',
+            headers: {
+                'csrf-token': csrfToken,
+                Accept: 'application/json',
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({
+                variables,
+                lists,
+            }),
         });
     }
 
@@ -2844,11 +2864,11 @@ Entry.VariableContainer = class VariableContainer {
                         content: JSON.stringify(func.content.toJSON()),
                     });
 
-                    blockList = func.content.getBlockList();
-                    const jsonData = this.getObjectVariables(blockList, findFuncKeys);
-                    functions = functions.concat(jsonData.functions);
-                    variables = variables.concat(jsonData.variables);
-                    messages = messages.concat(jsonData.messages);
+                    const contentBlockList = func.content.getBlockList();
+                    const jsonData = this.getObjectVariables(contentBlockList, findFuncKeys);
+                    functions = _.unionBy(functions, jsonData.functions, 'id');
+                    variables = _.unionBy(variables, jsonData.variables, 'id');
+                    messages = _.unionBy(messages, jsonData.messages, 'id');
                 }
             }
         });
@@ -2926,21 +2946,24 @@ Entry.VariableContainer = class VariableContainer {
     _maxNameLength = 10;
 
     clear() {
+        this.select(null);
         const _removeFunc = _.partial(_.result, _, 'remove');
-        const _destroyFunc = _.partial(_.result, _, 'destroy');
-
         const { engine = {}, container = {}, playground } = Entry;
 
         [...this.variables_, ...this.lists_].forEach(_removeFunc);
         _removeFunc(engine.projectTimer);
         _removeFunc(container.inputValue);
-        _.each(this.functions_, _destroyFunc);
+        _.each(this.functions_, this.removeFunction.bind(this));
 
         this.viewMode_ = 'all';
         this.variables_ = [];
         this.lists_ = [];
         this.messages_ = [];
         this.functions_ = {};
+
+        this._variableRefs = [];
+        this._messageRefs = [];
+        this._functionRefs = [];
 
         playground.reloadPlayground();
         this.updateList();
