@@ -221,11 +221,12 @@ module.exports = {
                         let [start, end] = script.getValues(['START', 'END'], script);
                         start = Number(start) * 1000;
                         end = Number(end) * 1000;
-
-                        Entry.Utils.playSound(sound.id, {
+                        const instance = Entry.Utils.playSound(sound.id, {
                             startTime: Math.min(start, end),
                             duration: Math.max(start, end) - Math.min(start, end),
                         });
+                        Entry.Utils.addSoundInstances(instance);
+
                     }
                     return script.callReturn();
                 },
@@ -375,6 +376,7 @@ module.exports = {
                         if (sound) {
                             script.playState = 1;
                             const instance = Entry.Utils.playSound(sound.id);
+                            Entry.Utils.addSoundInstances(instance);
                             const timeValue = script.getNumberValue('SECOND', script);
                             setTimeout(function() {
                                 instance.stop();
@@ -484,11 +486,12 @@ module.exports = {
                             const startValue = Math.min(start, end);
                             const endValue = Math.max(start, end);
                             const duration = endValue - startValue;
-
-                            Entry.Utils.playSound(sound.id, {
+                            const instance = Entry.Utils.playSound(sound.id, {
                                 startTime: startValue,
                                 duration,
                             });
+                            Entry.Utils.addSoundInstances(instance);
+
 
                             setTimeout(function() {
                                 script.playState = 0;
