@@ -9,9 +9,9 @@ Entry.Arduino = {
         ko: '아두이노',
         en: 'Arduino',
     },
-    setZero: function() {
+    setZero() {
         Entry.hw.sendQueue.readablePorts = [];
-        for (var port = 0; port < 20; port++) {
+        for (let port = 0; port < 20; port++) {
             Entry.hw.sendQueue[port] = 0;
             Entry.hw.sendQueue.readablePorts.push(port);
         }
@@ -23,92 +23,92 @@ Entry.Arduino = {
         height: 434,
         listPorts: {
             '2': {
-                name: Lang.Hw.port_en + ' 2 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 2 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '3': {
-                name: Lang.Hw.port_en + ' 3 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 3 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '4': {
-                name: Lang.Hw.port_en + ' 4 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 4 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '5': {
-                name: Lang.Hw.port_en + ' 5 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 5 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '6': {
-                name: Lang.Hw.port_en + ' 6 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 6 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '7': {
-                name: Lang.Hw.port_en + ' 7 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 7 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '8': {
-                name: Lang.Hw.port_en + ' 8 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 8 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '9': {
-                name: Lang.Hw.port_en + ' 9 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 9 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '10': {
-                name: Lang.Hw.port_en + ' 10 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 10 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '11': {
-                name: Lang.Hw.port_en + ' 11 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 11 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '12': {
-                name: Lang.Hw.port_en + ' 12 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 12 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             '13': {
-                name: Lang.Hw.port_en + ' 13 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} 13 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             a0: {
-                name: Lang.Hw.port_en + ' A0 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} A0 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             a1: {
-                name: Lang.Hw.port_en + ' A1 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} A1 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             a2: {
-                name: Lang.Hw.port_en + ' A2 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} A2 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             a3: {
-                name: Lang.Hw.port_en + ' A3 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} A3 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             a4: {
-                name: Lang.Hw.port_en + ' A4 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} A4 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
             a5: {
-                name: Lang.Hw.port_en + ' A5 ' + Lang.Hw.port_ko,
+                name: `${Lang.Hw.port_en} A5 ${Lang.Hw.port_ko}`,
                 type: 'input',
                 pos: { x: 0, y: 0 },
             },
@@ -144,7 +144,7 @@ Entry.Arduino.getBlocks = function() {
             paramsKeyMap: {
                 NAME: 0,
             },
-            func: function(sprite, script) {
+            func(sprite, script) {
                 return script.getStringField('NAME');
             },
             syntax: {
@@ -162,89 +162,6 @@ Entry.Arduino.getBlocks = function() {
                         keyOption: 'arduino_text',
                     },
                 ],
-            },
-        },
-        arduino_send: {
-            color: EntryStatic.colorSet.block.default.HARDWARE,
-            outerLine: EntryStatic.colorSet.block.darken.HARDWARE,
-            skeleton: 'basic',
-            statements: [],
-            params: [
-                {
-                    type: 'Block',
-                    accept: 'string',
-                },
-            ],
-            events: {},
-            def: {
-                params: [],
-            },
-            paramsKeyMap: {
-                VALUE: 0,
-            },
-            func: function(sprite, script) {
-                var signal = script.getValue('VALUE', script);
-                var xmlHttp = new XMLHttpRequest();
-                xmlHttp.open('POST', 'http://localhost:23518/arduino/', false);
-                xmlHttp.send(String(signal));
-                Entry.assert(xmlHttp.status == 200, 'arduino is not connected');
-                return script.callReturn();
-            },
-        },
-        arduino_get_number: {
-            color: EntryStatic.colorSet.block.default.HARDWARE,
-            outerLine: EntryStatic.colorSet.block.darken.HARDWARE,
-            skeleton: 'basic_string_field',
-            statements: [],
-            params: [
-                {
-                    type: 'Block',
-                    accept: 'string',
-                },
-            ],
-            events: {},
-            def: {
-                params: [],
-            },
-            paramsKeyMap: {
-                VALUE: 0,
-            },
-            func: function(sprite, script) {
-                var signal = script.getValue('VALUE', script);
-                var xmlHttp = new XMLHttpRequest();
-                xmlHttp.open('POST', 'http://localhost:23518/arduino/', false);
-                xmlHttp.send(String(signal));
-                Entry.assert(xmlHttp.status == 200, 'arduino is not connected');
-                var data = xmlHttp.responseText;
-                return Number(data);
-            },
-        },
-        arduino_get_string: {
-            color: EntryStatic.colorSet.block.default.HARDWARE,
-            outerLine: EntryStatic.colorSet.block.darken.HARDWARE,
-            skeleton: 'basic_string_field',
-            statements: [],
-            params: [
-                {
-                    type: 'Block',
-                    accept: 'string',
-                },
-            ],
-            events: {},
-            def: {
-                params: [],
-            },
-            paramsKeyMap: {
-                VALUE: 0,
-            },
-            func: function(sprite, script) {
-                var signal = script.getValue('VALUE', script);
-                var xmlHttp = new XMLHttpRequest();
-                xmlHttp.open('POST', 'http://localhost:23518/arduino/', false);
-                xmlHttp.send(String(signal));
-                Entry.assert(xmlHttp.status == 200, 'arduino is not connected');
-                var data = xmlHttp.responseText;
-                return data;
             },
         },
         arduino_get_sensor_number: {
@@ -276,7 +193,7 @@ Entry.Arduino.getBlocks = function() {
             paramsKeyMap: {
                 PORT: 0,
             },
-            func: function(sprite, script) {
+            func(sprite, script) {
                 return script.getStringField('PORT');
             },
             syntax: {
@@ -345,7 +262,7 @@ Entry.Arduino.getBlocks = function() {
             paramsKeyMap: {
                 PORT: 0,
             },
-            func: function(sprite, script) {
+            func(sprite, script) {
                 return script.getStringField('PORT');
             },
             syntax: {
@@ -412,7 +329,7 @@ Entry.Arduino.getBlocks = function() {
             paramsKeyMap: {
                 PORT: 0,
             },
-            func: function(sprite, script) {
+            func(sprite, script) {
                 return script.getStringField('PORT');
             },
             syntax: {
@@ -470,8 +387,8 @@ Entry.Arduino.getBlocks = function() {
             },
             class: 'arduino_value',
             isNotFor: ['arduino'],
-            func: function(sprite, script) {
-                var signal = script.getValue('VALUE', script);
+            func(sprite, script) {
+                const signal = script.getValue('VALUE', script);
                 return Entry.hw.getAnalogPortValue(signal[1]);
             },
             syntax: {
@@ -517,14 +434,14 @@ Entry.Arduino.getBlocks = function() {
             },
             class: 'arduino_value',
             isNotFor: ['arduino'],
-            func: function(sprite, script) {
+            func(sprite, script) {
                 const { hwModule = {} } = Entry.hw;
                 const { name } = hwModule;
                 if (name === 'ArduinoExt') {
                     return Entry.block.arduino_ext_get_digital.func(sprite, script);
                 } else {
-                    var signal = script.getNumberValue('PORT', script);
-                    return Entry.hw.getDigitalPortValue(signal);
+                    const signal = script.getNumberValue('PORT', script);
+                    return Entry.hw.portData.DIGITAL[signal];
                 }
             },
             syntax: {
@@ -586,10 +503,10 @@ Entry.Arduino.getBlocks = function() {
             },
             class: 'arduino_set',
             isNotFor: ['arduino'],
-            func: function(sprite, script) {
-                var port = script.getNumberValue('VALUE');
-                var operator = script.getField('OPERATOR');
-                var value = operator == 'on' ? 255 : 0;
+            func(sprite, script) {
+                const port = script.getNumberValue('VALUE');
+                const operator = script.getField('OPERATOR');
+                const value = operator == 'on' ? 255 : 0;
                 Entry.hw.setDigitalPortValue(port, value);
                 return script.callReturn();
             },
@@ -662,9 +579,9 @@ Entry.Arduino.getBlocks = function() {
             },
             class: 'arduino_set',
             isNotFor: ['arduino'],
-            func: function(sprite, script) {
-                var port = script.getNumberValue('PORT');
-                var value = script.getNumberValue('VALUE');
+            func(sprite, script) {
+                const port = script.getNumberValue('PORT');
+                let value = script.getNumberValue('VALUE');
                 value = Math.round(value);
                 value = Math.max(value, 0);
                 value = Math.min(value, 255);
@@ -763,16 +680,16 @@ Entry.Arduino.getBlocks = function() {
             },
             class: 'arduino',
             isNotFor: ['arduino'],
-            func: function(sprite, script) {
-                var value1 = script.getNumberValue('VALUE1', script);
-                var value2 = script.getNumberValue('VALUE2', script);
-                var value3 = script.getNumberValue('VALUE3', script);
-                var value4 = script.getNumberValue('VALUE4', script);
-                var value5 = script.getNumberValue('VALUE5', script);
+            func(sprite, script) {
+                const value1 = script.getNumberValue('VALUE1', script);
+                let value2 = script.getNumberValue('VALUE2', script);
+                let value3 = script.getNumberValue('VALUE3', script);
+                let value4 = script.getNumberValue('VALUE4', script);
+                let value5 = script.getNumberValue('VALUE5', script);
 
-                var stringValue4 = script.getValue('VALUE4', script);
-                var stringValue5 = script.getValue('VALUE5', script);
-                var isFloat = false;
+                const stringValue4 = script.getValue('VALUE4', script);
+                const stringValue5 = script.getValue('VALUE5', script);
+                let isFloat = false;
 
                 if (
                     (Entry.Utils.isNumber(stringValue4) && stringValue4.indexOf('.') > -1) ||
@@ -781,7 +698,7 @@ Entry.Arduino.getBlocks = function() {
                     isFloat = true;
                 }
 
-                var result = value1;
+                let result = value1;
                 if (value2 > value3) {
                     var swap = value2;
                     value2 = value3;
