@@ -111,10 +111,10 @@ Entry.HyACT_Xylobot.setLanguage = function() {
         },
         en: {
             template: {
-                hyact_xylobot_list_aixs_1: 'First aixs',
-                hyact_xylobot_list_aixs_2: 'Second aixs',
-                hyact_xylobot_list_aixs_3: 'Third aixs',
-                hyact_xylobot_list_aixs_all: 'All aixs',
+                hyact_xylobot_list_aixs_1: 'First axis',
+                hyact_xylobot_list_aixs_2: 'Second axis',
+                hyact_xylobot_list_aixs_3: 'Third axis',
+                hyact_xylobot_list_aixs_all: 'All axis',
                 hyact_xylobot_list_octave_4_do: 'C',
                 hyact_xylobot_list_octave_4_re: 'D',
                 hyact_xylobot_list_octave_4_mi: 'E',
@@ -122,7 +122,7 @@ Entry.HyACT_Xylobot.setLanguage = function() {
                 hyact_xylobot_list_octave_4_sol: 'G',
                 hyact_xylobot_list_octave_4_la: 'A',
                 hyact_xylobot_list_octave_4_si: 'B',
-                hyact_xylobot_list_octave_5_do: 'A high C',
+                hyact_xylobot_list_octave_5_do: 'High C',
                 hyact_xylobot_list_movement_on: 'On',
                 hyact_xylobot_list_movement_twinkle: 'Twinkle',
                 hyact_xylobot_list_movement_standby: 'Stand by',
@@ -146,7 +146,7 @@ Entry.HyACT_Xylobot.setLanguage = function() {
                 hyact_xylobot_set_led_rgb_block: 'LED color setting to R : %1% G : %2% B : %3% %4',
                 hyact_xylobot_set_torque_toggle_block: '%1 torque %2 %3',
                 hyact_xylobot_set_position_single_block: '%1 \'s target position setting to %2º(degree) %3',
-                hyact_xylobot_set_position_many_block: 'Target position setting to first aixs : %1º(degree)  second axis : %2º(degree)  third aixs : %3º(degree) %4',
+                hyact_xylobot_set_position_many_block: 'Target position setting to first axis : %1º(degree)  second axis : %2º(degree)  third axis : %3º(degree) %4',
                 hyact_xylobot_set_speed_block: '%1 \'s speed setting to %2% %3',
                 hyact_xylobot_set_torque_block: '%1 \'s torque value setting to %2% %3',
                 hyact_xylobot_set_play_note_block: 'Play %1 %2',
