@@ -106,7 +106,7 @@ Entry.Scroller = class Scroller {
 
     createScrollBar() {
         const r = this.RADIUS;
-
+        const { common = {} } = EntryStatic.colorSet || {};
         this.svgGroup = this.board.svg.elem('g').attr({ class: 'boardScrollbar' });
 
         if (this._horizontal) {
@@ -114,7 +114,7 @@ Entry.Scroller = class Scroller {
                 height: this.SCROLL_WIDTH,
                 rx: r,
                 ry: r,
-                fill: '#aac5d5',
+                fill: common.SCROLL_BAR || '#aac5d5',
                 class: 'scrollbar horizontal',
             });
             this.hScrollbar.type = 'horizontal';
@@ -126,7 +126,7 @@ Entry.Scroller = class Scroller {
                 width: this.SCROLL_WIDTH,
                 rx: r,
                 ry: r,
-                fill: '#aac5d5',
+                fill: common.SCROLL_BAR || '#aac5d5',
                 class: 'scrollbar vertical',
             });
             this.vScrollbar.type = 'vertical';
