@@ -422,8 +422,9 @@ Entry.Workspace = class Workspace {
             }
             const mainWorksapceMode = Entry.playground.mainWorkspace.getMode();
             const playgroundMode = Entry.playground.getViewMode();
-            const isBlockCodeView = (mainWorksapceMode === Entry.Workspace.MODE_OVERLAYBOARD ||
-                mainWorksapceMode === Entry.Workspace.MODE_BOARD) &&
+            const isBlockCodeView =
+                (mainWorksapceMode === Entry.Workspace.MODE_OVERLAYBOARD ||
+                    mainWorksapceMode === Entry.Workspace.MODE_BOARD) &&
                 (playgroundMode === 'code' || playgroundMode === 'variable');
             switch (keyCode) {
                 case 86: //paste
@@ -586,8 +587,7 @@ Entry.Workspace = class Workspace {
 
         function checkObjectAndAlert(object, message) {
             if (!object) {
-                message =
-                    message || '오브젝트가 존재하지 않습니다. 오브젝트를 추가한 후 시도해주세요.';
+                message = message || Lang.Workspace.object_not_exist_error;
                 entrylms.alert(message);
                 return false;
             }
